% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb_prcp_qm.R
\name{apply_quantile_mapping}
\alias{apply_quantile_mapping}
\title{Apply quantile mapping transformation}
\usage{
apply_quantile_mapping(prcp, mon, year, base_gamma, target_gamma)
}
\description{
Apply quantile mapping transformation
}
\keyword{internal}
