% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSpellLength.R
\name{averageSpellLength}
\alias{averageSpellLength}
\title{Calculate length of a wet or a dry spell}
\usage{
averageSpellLength(x, threshold = 0, below = TRUE)
}
\arguments{
\item{x}{numeric vector of values}

\item{threshold}{numeric threshold to define the wet or a dry spell}

\item{below}{binary variable. TRUE considers values less than or equal, FALSE above the threshold.}
}
\description{
Calculate length of a wet or a dry spell
}
