% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchClimateSurface.R
\name{batchClimateSurface}
\alias{batchClimateSurface}
\title{Batch create climate response surfaces from CST Toolbox applications}
\usage{
batchClimateSurface(
  project.dir = NULL,
  str.data = NULL,
  gcm.data = NULL,
  gcm.legend = TRUE,
  gcm.future.period = "near",
  metrics = c("mean", "min", "max", "q95", "Q7day_max", "Q7day_min", "drymonth_mean",
    "wetmonth_mean"),
  metric.labels = c(bquote(bold(`Mean flow `(m^3/s))),
    bquote(bold(`Minimum daily flow `(m^3/s))),
    bquote(bold(`Maximum daily flow `(m^3/s))),
    bquote(bold(`95th quantile flow `(m^3/s))),
    bquote(bold(`7-Day maximum flow `(m^3/s))),
    bquote(bold(`7-Day minimum flow `(m^3/s))),
    bquote(bold(`Mean flow during driest month `(m^3/s))),
    bquote(bold(`Mean flow during wettest month `(m^3/s)))),
  metric.direction = c(1, 1, 0, 1, 0, 1, 1, 0),
  variable.x.breaks = NULL,
  variable.y.breaks = NULL,
  ...
)
}
\arguments{
\item{project.dir}{placeholder}

\item{str.data}{placeholder}

\item{gcm.data}{placeholder}

\item{gcm.future.period}{placeholder}

\item{metrics}{placeholder}

\item{metric.labels}{placeholder}

\item{metric.direction}{placeholder}

\item{variable.x.breaks}{placeholder}

\item{variable.y.breaks}{placeholder}
}
\description{
Batch create climate response surfaces from CST Toolbox applications
}
