% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCreateResponseSurfaces.R
\name{batchCreateResponseSurfaces}
\alias{batchCreateResponseSurfaces}
\title{Batch create climate response surfaces from CST Toolbox applications}
\usage{
batchCreateResponseSurfaces(
  str.data = NULL,
  gcm.data = NULL,
  project.dir = NULL,
  tavg.steps = NULL,
  prcp.steps = NULL,
  gcm.future.period = "near",
  metrics = c("mean", "min", "max", "q95", "Q7day_max", "Q7day_min", "drymonth_mean",
    "wetmonth_mean"),
  metric.labels = c(bquote(bold(`Mean flow `(m^3/s))),
    bquote(bold(`Minimum daily flow `(m^3/s))),
    bquote(bold(`Maximum daily flow `(m^3/s))),
    bquote(bold(`95th quantile flow `(m^3/s))),
    bquote(bold(`7-Day maximum flow `(m^3/s))),
    bquote(bold(`7-Day minimum flow `(m^3/s))),
    bquote(bold(`Mean flow during driest month `(m^3/s))),
    bquote(bold(`Mean flow during wettest month `(m^3/s)))),
  metric.direction = c(1, 1, 0, 1, 0, 1, 1, 0)
)
}
\arguments{
\item{str.data}{placeholder}

\item{gcm.data}{placeholder}

\item{project.dir}{placeholder}

\item{tavg.steps}{placeholder}

\item{prcp.steps}{placeholder}

\item{gcm.future.period}{placeholder}

\item{metrics}{placeholder}

\item{metric.labels}{placeholder}

\item{metric.direction}{placeholder}
}
\description{
Batch create climate response surfaces from CST Toolbox applications
}
