% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSpellLength.R
\name{calculateSpellLength}
\alias{calculateSpellLength}
\title{Calculate Spell Length}
\usage{
calculateSpellLength(x, threshold = 0, below = TRUE)
}
\arguments{
\item{x}{placeholder}

\item{threshold}{placeholder}

\item{below}{placeholder}
}
\value{

}
\description{
Calculate Spell Length
}
