% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateSurface.R
\name{climateSurface}
\alias{climateSurface}
\title{Climate Response Surface Generator}
\usage{
climateSurface(
  str.data = NULL,
  gcm.data = NULL,
  variable.x = NULL,
  variable.y = NULL,
  variable.z = NULL,
  threshold.z = NULL,
  plot.title = "change in variable",
  variable.x.label = "change in temperature",
  variable.y.label = "change in precipitation",
  failure.direction = "low",
  gcm.scenario.list = c("rcp26", "rcp45", "rcp60", "rcp85"),
  gcm.bivariate.dist = FALSE,
  gcm.transparency = 0.85,
  gcm.legend = TRUE,
  variable.z.min = NULL,
  variable.z.max = NULL,
  variable.z.bin = 15,
  variable.z.min.legend = NULL,
  variable.z.max.legend = NULL,
  variable.z.bin.legend = 11,
  variable.x.breaks = NULL,
  variable.y.breaks = NULL,
  text.scale = 0.8
)
}
\arguments{
\item{gcm.scenario.list}{}
}
\description{
Climate Response Surface Generator
}
