% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{compute_conditional_precip_cor}
\alias{compute_conditional_precip_cor}
\title{Conditional precip-X correlations within each grid}
\usage{
compute_conditional_precip_cor(
  data,
  variables,
  mc.thresholds = NULL,
  wet_def = c("gt0", "monthly_quantile"),
  use_anom = TRUE,
  use_log_precip_on_wet = TRUE,
  method = "pearson",
  min_pairs = 50
)
}
\description{
Conditional precip-X correlations within each grid
}
\keyword{internal}
