% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{compute_grouped_statistics}
\alias{compute_grouped_statistics}
\title{Compute grouped statistics}
\usage{
compute_grouped_statistics(df, variables, group.vars, stat.funs)
}
\description{
Compute grouped statistics
}
\keyword{internal}
