% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{compute_monthly_diagnostics}
\alias{compute_monthly_diagnostics}
\title{Compute month-by-month diagnostics for quantile-mapped precipitation}
\usage{
compute_monthly_diagnostics(
  prcp_ref,
  prcp_adj,
  month,
  mean_factor = NULL,
  var_factor = NULL,
  year = NULL
)
}
\description{
Compute month-by-month diagnostics for quantile-mapped precipitation
}
\keyword{internal}
