% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quantile_mapping.R
\name{compute_quantile_diagnostics}
\alias{compute_quantile_diagnostics}
\title{Compute quantile diagnostics}
\usage{
compute_quantile_diagnostics(value.original, value.adjusted, quantiles)
}
\description{
Compute quantile diagnostics
}
\keyword{internal}
