% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{compute_realization_fit_metrics}
\alias{compute_realization_fit_metrics}
\title{Compute fit metrics for each realization}
\usage{
compute_realization_fit_metrics(obs.results, sim.results, variables)
}
\description{
Compute fit metrics for each realization
}
\keyword{internal}
