% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quantile_mapping.R
\name{compute_spell_diagnostics}
\alias{compute_spell_diagnostics}
\title{Compute spell length diagnostics}
\usage{
compute_spell_diagnostics(value.original, value.adjusted, threshold)
}
\description{
Compute spell length diagnostics
}
\keyword{internal}
