% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{compute_summary_metrics}
\alias{compute_summary_metrics}
\title{Compute overall summary metrics for quantile-mapping diagnostics}
\usage{
compute_summary_metrics(
  moments,
  quant_metrics,
  extreme_metrics,
  spell_metrics,
  dryday_metrics
)
}
\description{
Compute overall summary metrics for quantile-mapping diagnostics
}
\keyword{internal}
