% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_diagnostics.R
\name{compute_temporal_diagnostics}
\alias{compute_temporal_diagnostics}
\title{Compute temporal diagnostics (no subjective categories)}
\usage{
compute_temporal_diagnostics(precip_ref, precip_adj, month, year, mask)
}
\description{
Compute temporal diagnostics (no subjective categories)
}
\keyword{internal}
