% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator_plots.R
\name{create_daily_mean_plot}
\alias{create_daily_mean_plot}
\title{Create daily mean plot}
\usage{
create_daily_mean_plot(
  daily.stats.season,
  plot.config,
  show.title,
  save.plots,
  output.path
)
}
\description{
Create daily mean plot
}
\keyword{internal}
