% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator_plots.R
\name{create_intergrid_cor_plot}
\alias{create_intergrid_cor_plot}
\title{Create inter-grid correlation plot}
\usage{
create_intergrid_cor_plot(
  stats.intercor,
  plot.config,
  show.title,
  save.plots,
  output.path
)
}
\description{
Create inter-grid correlation plot
}
\keyword{internal}
