% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{create_monthly_pattern_plot}
\alias{create_monthly_pattern_plot}
\title{Create monthly pattern plot for a single variable}
\usage{
create_monthly_pattern_plot(
  stats.mon.aavg.sim,
  stats.mon.aavg.obs,
  variable,
  variable.label,
  plot.config,
  show.title,
  save.plots,
  output.path
)
}
\description{
Create monthly pattern plot for a single variable
}
\keyword{internal}
