% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{create_spell_length_plot}
\alias{create_spell_length_plot}
\title{Create spell length plot}
\usage{
create_spell_length_plot(
  stats.wetdry,
  plot.config,
  show.title,
  save.plots,
  output.path
)
}
\description{
Create spell length plot
}
\keyword{internal}
