% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator_plots.R
\name{create_wetdry_days_plot}
\alias{create_wetdry_days_plot}
\title{Create wet/dry days count plot}
\usage{
create_wetdry_days_plot(
  stats.wetdry,
  plot.config,
  show.title,
  save.plots,
  output.path
)
}
\description{
Create wet/dry days count plot
}
\keyword{internal}
