% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyPerformance.R
\name{dailyPerformance}
\alias{dailyPerformance}
\title{Function to assess the weather generator skill}
\usage{
dailyPerformance(
  daily.sim = NULL,
  daily.obs = NULL,
  out.path = NULL,
  variables = NULL,
  variable.labels = NULL,
  variable.units = NULL
)
}
\arguments{
\item{daily.sim}{A matrix of daily time-series of weather variables}

\item{daily.obs}{A vector of daily time-series of observed weather series}

\item{out.path}{A character string to define the path of resulting netcdf file.}

\item{variables}{A vector object specifying the names of weather variables.}

\item{variable.labels}{A vector object specifying the names of weather variable labels.}

\item{variable.units}{A vector object specifying the names of weather variable units.}
}
\value{

}
\description{
Function to assess the weather generator skill
}
