% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator_plots.R
\name{.add_line_if_possible}
\alias{.add_line_if_possible}
\title{Add geom_line only for groups with >= 2 observations}
\usage{
.add_line_if_possible(p, data, mapping, group_col, ...)
}
\description{
Add geom_line only for groups with >= 2 observations
}
\keyword{internal}
