% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_modwt.R
\name{.as_pos_int_scalar}
\alias{.as_pos_int_scalar}
\title{Validate a positive integer scalar argument}
\usage{
.as_pos_int_scalar(x, name)
}
\arguments{
\item{x}{Object. Candidate scalar.}

\item{name}{Character scalar. Argument name.}
}
\value{
Integer scalar.
}
\description{
Validate a positive integer scalar argument
}
\keyword{internal}
