% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_warm.R
\name{.block_bootstrap}
\alias{.block_bootstrap}
\title{Block bootstrap for one realization}
\usage{
.block_bootstrap(x, n, block_len)
}
\arguments{
\item{x}{Numeric vector.}

\item{n}{Integer. Output length.}

\item{block_len}{Integer. Block length.}
}
\value{
Numeric vector length n.
}
\description{
Block bootstrap for one realization
}
\keyword{internal}
