% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{.build_plot_data}
\alias{.build_plot_data}
\title{Build data for plotting}
\usage{
.build_plot_data(obs_results, sim_results, variables)
}
\arguments{
\item{obs_results}{Output from `.summarize_observed_data()`.}

\item{sim_results}{Output from `.summarize_simulated_data()`.}

\item{variables}{Character vector of variable names.}
}
\value{
List of standardized data frames for diagnostic plots.
}
\description{
Build data for plotting
}
\keyword{internal}
