% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_modwt.R
\name{.check_boundary}
\alias{.check_boundary}
\title{Validate boundary handling option}
\usage{
.check_boundary(boundary)
}
\arguments{
\item{boundary}{Character scalar. Boundary option.}
}
\value{
Character scalar. Normalized boundary.
}
\description{
Validate boundary handling option
}
\keyword{internal}
