% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_diagnostics.R
\name{.compute_intended_ratios}
\alias{.compute_intended_ratios}
\title{Compute intended scaling ratios over the evaluated subset (time-weighted)}
\usage{
.compute_intended_ratios(mask, month, year, mean_factor, var_factor)
}
\description{
Compute intended scaling ratios over the evaluated subset (time-weighted)
}
\keyword{internal}
