% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_warm.R
\name{.default_block_len}
\alias{.default_block_len}
\title{Default block length for block bootstrap}
\usage{
.default_block_len(n)
}
\arguments{
\item{n}{Integer. Series length.}
}
\value{
Integer block length.
}
\description{
Default block length for block bootstrap
}
\keyword{internal}
