% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_warm.R
\name{.fit_warm_arima_forecast}
\alias{.fit_warm_arima_forecast}
\title{Fit ARIMA for bypass mode}
\usage{
.fit_warm_arima_forecast(
  x,
  max_p,
  max_q,
  stationary = TRUE,
  include_mean = FALSE,
  allow_drift = FALSE
)
}
\arguments{
\item{x}{Numeric vector (centered).}

\item{max_p}{Integer. Maximum AR order.}

\item{max_q}{Integer. Maximum MA order.}

\item{stationary}{Logical. Whether to enforce stationarity.}

\item{include_mean}{Logical. Whether to include mean.}

\item{allow_drift}{Logical. Whether to allow drift.}
}
\value{
List with model or parameter entries, or NULL.
}
\description{
Fit ARIMA for bypass mode
}
\keyword{internal}
