% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_modwt.R
\name{.get_filter_length}
\alias{.get_filter_length}
\title{Get Wavelet Filter Length}
\usage{
.get_filter_length(filter)
}
\arguments{
\item{filter}{Character. Filter name.}
}
\value{
Integer. Filter length.
}
\description{
Returns the filter length for a named wavelet filter.
}
\keyword{internal}
