% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_diagnostics.R
\name{.make_drywet_table}
\alias{.make_drywet_table}
\title{Build dry/wet day count table for printing}
\usage{
.make_drywet_table(drydays)
}
\description{
Build dry/wet day count table for printing
}
\keyword{internal}
