% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_diagnostics.R
\name{.make_extremes_table}
\alias{.make_extremes_table}
\title{Build extremes table for printing}
\usage{
.make_extremes_table(extremes)
}
\description{
Build extremes table for printing
}
\keyword{internal}
