% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_diagnostics.R
\name{.make_quantiles_table}
\alias{.make_quantiles_table}
\title{Build quantiles table for printing}
\usage{
.make_quantiles_table(quantiles)
}
\description{
Build quantiles table for printing
}
\keyword{internal}
