% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_diagnostics.R
\name{.make_spells_table}
\alias{.make_spells_table}
\title{Build spell lengths table for printing}
\usage{
.make_spells_table(spells)
}
\description{
Build spell lengths table for printing
}
\keyword{internal}
