% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet.R
\name{.pet_monthly_hargreaves}
\alias{.pet_monthly_hargreaves}
\title{Monthly PET via Hargreaves (FAO-56)}
\usage{
.pet_monthly_hargreaves(month, temp, temp_range, lat_deg)
}
\description{
Monthly PET via Hargreaves (FAO-56)
}
\keyword{internal}
