% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{.pick_contiguous_year_window}
\alias{.pick_contiguous_year_window}
\title{Pick a random contiguous window of years of length `window_years` from a year vector}
\usage{
.pick_contiguous_year_window(years, window_years)
}
\arguments{
\item{years}{Integer vector of available years.}

\item{window_years}{Integer number of contiguous years to sample.}
}
\value{
Integer vector of selected years in chronological order.
}
\description{
Pick a random contiguous window of years of length `window_years` from a year vector
}
\keyword{internal}
