% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{.summarize_observed_data}
\alias{.summarize_observed_data}
\title{Summarize observed weather data}
\usage{
.summarize_observed_data(
  daily_obs,
  variables,
  grid_count,
  wet_quantile,
  extreme_quantile
)
}
\description{
Summarize observed weather data
}
\keyword{internal}
