% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{.summarize_simulated_data}
\alias{.summarize_simulated_data}
\title{Summarize simulated weather data}
\usage{
.summarize_simulated_data(daily_sim, n_realizations, variables, mc_thresholds)
}
\description{
Summarize simulated weather data
}
\keyword{internal}
