% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{.validate_assessment_inputs}
\alias{.validate_assessment_inputs}
\title{Validate inputs for weather assessment}
\usage{
.validate_assessment_inputs(
  daily_sim,
  daily_obs,
  variables,
  n_realizations,
  wet_quantile,
  extreme_quantile,
  verbose
)
}
\description{
Validate inputs for weather assessment
}
\keyword{internal}
