% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_warm.R
\name{.warm_arima_viable}
\alias{.warm_arima_viable}
\title{Check ARIMA viability for WARM components}
\usage{
.warm_arima_viable(x, min_n = 8L, sd_eps = 1e-08, min_unique = 3L)
}
\arguments{
\item{x}{Numeric vector.}

\item{min_n}{Integer. Minimum length.}

\item{sd_eps}{Numeric. Minimum standard deviation threshold.}

\item{min_unique}{Integer. Minimum unique values.}
}
\value{
Logical scalar.
}
\description{
Check ARIMA viability for WARM components
}
\keyword{internal}
