% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_warm.R
\name{.warm_block_bootstrap}
\alias{.warm_block_bootstrap}
\title{Block bootstrap for WARM components}
\usage{
.warm_block_bootstrap(x, n, block_len = NULL)
}
\arguments{
\item{x}{Numeric vector.}

\item{n}{Integer. Output length.}

\item{block_len}{Integer. Block length.}
}
\value{
Numeric vector length n.
}
\description{
Block bootstrap for WARM components
}
\keyword{internal}
