% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_warm.R
\name{.warm_simulate_from_fit}
\alias{.warm_simulate_from_fit}
\title{Simulate from ARIMA fit for WARM}
\usage{
.warm_simulate_from_fit(fit, n, n_sim)
}
\arguments{
\item{fit}{List. Output from .warm_fit_arima_safe or .fit_warm_arima_forecast.}

\item{n}{Integer. Simulation length.}

\item{n_sim}{Integer. Number of realizations.}
}
\value{
Numeric matrix n by n_sim.
}
\description{
Simulate from ARIMA fit for WARM
}
\keyword{internal}
