% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator_plots.R
\name{export_multipanel_plot}
\alias{export_multipanel_plot}
\title{Export multi-panel plot}
\usage{
export_multipanel_plot(
  p,
  filename,
  show.title,
  save.plots,
  title = NULL,
  subtitle = NULL,
  output.path
)
}
\description{
Export multi-panel plot
}
\keyword{internal}
