% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{filter_df_to_years}
\alias{filter_df_to_years}
\title{Filter a single grid data.frame to selected years}
\usage{
filter_df_to_years(df, year_keep)
}
\description{
Filter a single grid data.frame to selected years
}
\keyword{internal}
