% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{find_leap_day_indices}
\alias{find_leap_day_indices}
\title{Find leap-day (Feb 29) indices in a date vector}
\usage{
find_leap_day_indices(date)
}
\arguments{
\item{date}{A vector coercible to \code{Date}.}
}
\value{
Integer vector of indices where dates equal February 29, or \code{NULL} if none.
}
\description{
Scans a vector of dates and returns the integer indices of all elements that
correspond to February 29. If the vector contains no leap days, returns \code{NULL}.
}
\examples{
find_leap_day_indices(as.Date(c("1980-02-28", "1980-02-29", "1981-01-01")))
find_leap_day_indices(as.Date(c("2001-01-01", "2001-12-31")))

}
