% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_helpers.R
\name{find_leap_days}
\alias{find_leap_days}
\title{Find leap-day positions (Feb 29) in a date vector}
\usage{
find_leap_days(dates)
}
\arguments{
\item{dates}{A vector coercible to \code{Date}.}
}
\value{
Integer vector of indices where dates equal February 29, or
  \code{NULL} if no leap days are present.
}
\description{
Scans a vector of dates and returns the integer indices of all elements that
correspond to February 29. If the vector contains no leap days, the function
returns \code{NULL}.
}
\examples{
find_leap_days(as.Date(c("1980-02-28", "1980-02-29", "1981-01-01")))
# Returns: 2

find_leap_days(as.Date(c("2001-01-01", "2001-12-31")))
# Returns: NULL

}
