% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_warm_logging.R
\name{fmt_num}
\alias{fmt_num}
\title{Format numeric with specified digits}
\usage{
fmt_num(x, digits = 4L)
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Number of decimal places}
}
\value{
Character string
}
\description{
Format numeric with specified digits
}
\keyword{internal}
