% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warm_filtering.R
\name{fmt_pct}
\alias{fmt_pct}
\title{Format percentage}
\usage{
fmt_pct(x, digits = 1L)
}
\arguments{
\item{x}{Numeric proportion (0 to 1)}

\item{digits}{Number of decimal places}
}
\value{
Character string with percent sign
}
\description{
Format percentage
}
\keyword{internal}
