% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcmPlausibility.R
\name{GCMplausiblity}
\alias{GCMplausiblity}
\title{Calculate likelihood from gcm projections}
\usage{
GCMplausiblity(
  str.data = NULL,
  gcm.data = NULL,
  gcm.scenario.list = c("ssp126", "ssp245", "ssp370", "ssp585"),
  clevel.list = c(0.5, 0.95),
  metric.list = NULL,
  metric.labs = NULL,
  location.list = NULL
)
}
\arguments{
\item{str.data}{placeholder}

\item{gcm.data}{placeholder}

\item{gcm.scenario.list}{placeholder}

\item{clevel.list}{placeholder}

\item{metric.list}{placeholder}

\item{metric.labs}{placeholder}

\item{location.list}{placeholder}
}
\description{
Calculate likelihood from gcm projections
}
