% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDailyWeather.R
\name{generateDailyWeather}
\alias{generateDailyWeather}
\title{Main wrapper function for daily weather generation}
\usage{
generateDailyWeather(
  project_name = NULL,
  wg_date_begin = NULL,
  wg_date_end = NULL,
  sim_date_begin = NULL,
  wg_variables = NULL,
  wg_variable_labs = NULL,
  wg_variable_units = NULL,
  warm_variable = NULL,
  ssig = 0.9,
  ymax = 40,
  rmax = 10000,
  nmax = 5,
  PARCC = NULL,
  nc.path = NULL,
  nc.file = NULL,
  nc.dimnames = NULL,
  validate = FALSE
)
}
\arguments{
\item{project_name}{Placeholder.}

\item{wg_date_begin}{Placeholder.}

\item{wg_date_end}{Placeholder.}

\item{sim_date_begin}{Placeholder.}

\item{wg_variables}{Placeholder.}

\item{wg_variable_labs}{Placeholder.}

\item{wg_variable_units}{Placeholder.}

\item{warm_variable}{Placeholder.}

\item{ssig}{Placeholder.}

\item{ymax}{Placeholder.}

\item{rmax}{Placeholder.}

\item{nmax}{Placeholder.}

\item{PARCC}{Placeholder.}

\item{nc.path}{Placeholder.}

\item{nc.file}{Placeholder.}

\item{nc.dimnames}{Placeholder.}

\item{validate}{Placeholder.}

\item{out_path}{Placeholder.}
}
\value{

}
\description{
Main wrapper function for daily weather generation
}
