% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateWeatherSeries.R
\name{generateWeatherSeries}
\alias{generateWeatherSeries}
\title{Simulate gridded weather function}
\usage{
generateWeatherSeries(
  weather.data = NULL,
  weather.grid = NULL,
  weather.date = NULL,
  variable.names = NULL,
  variable.labels = NULL,
  variable.units = NULL,
  sim.year.num = NULL,
  sim.year.start = 2020,
  month.start = 1,
  realization.num = 5,
  warm.variable = "precip",
  warm.signif.level = 0.9,
  warm.sample.num = 5000,
  warm.subset.criteria = list(mean = c(0.9, 1.1), sd = c(0.8, 1.2), min = c(0.6, 1.3),
    max = c(0.7, 1.4), power = c(0.2, 10), nonsignif.threshold = 1.2),
  knn.sample.num = 120,
  mc.wet.quantile = 0.3,
  mc.extreme.quantile = 0.8,
  dry.spell.change = rep(1, 12),
  wet.spell.change = rep(1, 12),
  output.path = tempdir(),
  seed = NULL,
  compute.parallel = FALSE,
  num.cores = NULL,
  save.rdata = FALSE
)
}
\arguments{
\item{weather.data}{list of data frames of daily weather observations per grid cell. Each data frame, columns are weather variables and rows are daily values.}

\item{weather.grid}{Data frame of grid cells. Each grid cell is assigned an id starting from 1, x and y coordinate index value, and x and y coordinates.}

\item{weather.date}{a vector of dates matching the weather.data}

\item{variable.names}{vector of names for the variables to be included in the weather generator}

\item{variable.labels}{vector of labels for the weather variables (optional). If no values provided, it is labels will be same as the names}

\item{variable.units}{vector of units for each of the weather variables (optional). If no values provided, a blank vector is used.}

\item{sim.year.num}{numeric value indicating the desired total number of years of simulated weather realizations}

\item{sim.year.start}{numeric value indicating the starting year of the generated time-series}

\item{month.start}{the first month of the water year (default value is 1).}

\item{realization.num}{number of natural variability realizations to be generated.}

\item{warm.variable}{the name of the variable for the wavelet auto regressive mode. Default value is precipitation variable.}

\item{warm.signif.level}{the significance level for the warm model.}

\item{warm.sample.num}{number of annual sequences to be generated from the the warm model}

\item{warm.subset.criteria}{A list of statistical parameters used for subsetting from the initial annual simulated series}

\item{knn.sample.num}{number of knn years to be sampled}

\item{mc.wet.quantile}{wet state threshold (quantile value) for markov-chain modeling}

\item{mc.extreme.quantile}{extremely wet state threshold (quantile value) for markov-chain modeling}

\item{dry.spell.change}{placeholder}

\item{wet.spell.change}{placeholder}

\item{output.path}{output path for the weather generator results (string)}

\item{seed}{a random seed value (numeric)}

\item{compute.parallel}{logical value indicating whether to run (some) functions in parallel}

\item{num.cores}{Number of cores to be allocated for parallel computing. If left NULL, maximum possible cores minus one is assigned}
}
\value{
a list object with two elements
 \item{resampled}{data frame of the sampled dates for each synthetic trace}
 \item{dates}{a date vector corresponding to simulated dates}
}
\description{
A short description...
}
\details{
Additional details...
}
