% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{get_full_year_block}
\alias{get_full_year_block}
\title{Find full (365-day) years after leap-day removal and return longest contiguous block}
\usage{
get_full_year_block(date)
}
\description{
Find full (365-day) years after leap-day removal and return longest contiguous block
}
\keyword{internal}
