% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPI.R
\name{getPI}
\alias{getPI}
\title{Compute Stationary Distribution of a 3-State Markov Chain}
\usage{
getPI(p00, p01, p02, p10, p11, p12, p20, p21, p22)
}
\arguments{
\item{p00}{Probability of transitioning from state 0 to 0}

\item{p01}{Probability of transitioning from state 0 to 1}

\item{p02}{Probability of transitioning from state 0 to 2}

\item{p10}{Probability of transitioning from state 1 to 0}

\item{p11}{Probability of transitioning from state 1 to 1}

\item{p12}{Probability of transitioning from state 1 to 2}

\item{p20}{Probability of transitioning from state 2 to 0}

\item{p21}{Probability of transitioning from state 2 to 1}

\item{p22}{Probability of transitioning from state 2 to 2}
}
\value{
A numeric vector of stationary probabilities.
}
\description{
Solves for the stationary probabilities of a 3-state Markov chain
given the transition probabilities between all states.
}
\keyword{internal}
