% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWaterYear.R
\name{getWaterYear}
\alias{getWaterYear}
\title{Water year function}
\usage{
getWaterYear(date = NULL, month.list = 1:12)
}
\arguments{
\item{date}{placeholder}

\item{month.list}{placeholder}
}
\value{

}
\description{
Water year function
}
