% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hargreavesPET.R
\name{hargreavesPET}
\alias{hargreavesPET}
\title{Potential evapotranspiration (PET) by hargreaves method}
\usage{
hargreavesPET(months, temp, tdiff, lat)
}
\arguments{
\item{months}{a time-series date object}

\item{temp}{a vector of average temperature values (°C)}

\item{lat}{latitude information (negative values for southern hemisphere)}

\item{tdif}{a vector of differences computed from maximum and minimum
temperatures}
}
\value{
the output is a vector of PET values
}
\description{
\url{http://www.civil.uwaterloo.ca/watflood/manual/02_03_2.htm}
}
