% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeClimateChanges.R
\name{imposeClimateChanges}
\alias{imposeClimateChanges}
\title{Perturb weather realizations}
\usage{
imposeClimateChanges(
  climate.data = NULL,
  climate.grid = NULL,
  sim.dates = NULL,
  change.factor.precip.mean = NULL,
  change.factor.precip.variance = NULL,
  change.factor.temp.mean = NULL,
  change.type.temp = "transient",
  change.type.precip = "transient",
  calculate.pet = TRUE
)
}
\arguments{
\item{climate.data}{placeholder}

\item{climate.grid}{placeholder}

\item{sim.dates}{placeholder}

\item{change.factor.precip.mean}{placeholder}

\item{change.factor.precip.variance}{placeholder}

\item{change.factor.temp.mean}{placeholder}

\item{change.type.temp}{placeholder}

\item{change.type.precip}{placeholder}

\item{calculate.pet}{placeholder}
}
\value{

}
\description{
Perturb weather realizations
}
