% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNearestNeighbors.R
\name{kNearestNeighbors}
\alias{kNearestNeighbors}
\title{Implementation of K Nearest Neighbor Sampling}
\usage{
kNearestNeighbors(x = NULL, k = 5, s = NULL, seed = NULL)
}
\arguments{
\item{x}{A numeric vector of time-series of weather variables xt, t=1,..,n}

\item{k}{A numeric value to set the number of neighbors to resample from}

\item{s}{A numeric value for calculating the euclidean distance vector}
}
\value{
numeric, resampled value from the time-series vector x
}
\description{
This function applies a simple K nearest neighbors sampling scheme to find
the historical nearest neighbors of the current variable value and to resample
from k-nearest values. Adopted from Rajagopalan and Lall (1999)
}
