% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_ANNUAL.R
\name{KNN_ANNUAL}
\alias{KNN_ANNUAL}
\title{This KNN function generates 100 years.....}
\usage{
KNN_ANNUAL(
  sim_annual_prcp,
  ANNUAL_PRCP,
  WATER_YEAR_A,
  kk,
  k1,
  y,
  y_sample_size = 20
)
}
\arguments{
\item{sim_annual_prcp}{placeholder}

\item{ANNUAL_PRCP}{placeholder}

\item{WATER_YEAR_A}{placeholder}

\item{kk}{placeholder}

\item{k1}{placeholder}

\item{y}{placeholder}

\item{y_sample_size}{placeholder}
}
\value{

}
\description{
This KNN function generates 100 years.....
}
