% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnAnnual.R
\name{knnAnnual}
\alias{knnAnnual}
\title{This KNN function generates 100 years.....}
\usage{
knnAnnual(
  sim_annual_prcp = NULL,
  ANNUAL_PRCP = NULL,
  WATER_YEAR_A = NULL,
  kk = NULL,
  k1 = NULL,
  y = NULL,
  y_sample_size = 20
)
}
\arguments{
\item{sim_annual_prcp}{placeholder}

\item{ANNUAL_PRCP}{placeholder}

\item{WATER_YEAR_A}{placeholder}

\item{kk}{placeholder}

\item{k1}{placeholder}

\item{y}{placeholder}

\item{y_sample_size}{placeholder}
}
\value{

}
\description{
This KNN function generates 100 years.....
}
