% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnDaily.R
\name{knnDaily}
\alias{knnDaily}
\title{KNN daily sampling}
\usage{
knnDaily(
  cur_sim_PRCP,
  cur_sim_TEMP,
  PRCP_TODAY,
  TEMP_TODAY,
  DATE_TOMORROW,
  k,
  sd_monthly_PRCP,
  sd_monthly_TEMP,
  mean_monthly_PRCP,
  mean_monthly_TEMP,
  k1,
  count
)
}
\arguments{
\item{cur_sim_PRCP}{placeholder}

\item{cur_sim_TEMP}{placeholder}

\item{PRCP_TODAY}{placeholder}

\item{TEMP_TODAY}{placeholder}

\item{DATE_TOMORROW}{placeholder}

\item{k}{placeholder}

\item{sd_monthly_PRCP}{placeholder}

\item{sd_monthly_TEMP}{placeholder}

\item{mean_monthly_PRCP}{placeholder}

\item{mean_monthly_TEMP}{placeholder}

\item{k1}{placeholder}

\item{count}{placeholder}
}
\value{

}
\description{
KNN daily sampling
}
