% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warm_filtering.R
\name{log_filtering_setup}
\alias{log_filtering_setup}
\title{Log initial setup information}
\usage{
log_filtering_setup(n_obs, n_sim, n_realizations, sample_num, relax_priority)
}
\arguments{
\item{n_obs}{Number of observations in observed series}

\item{n_sim}{Number of years in simulated series}

\item{n_realizations}{Number of realizations}

\item{sample_num}{Target number to sample}

\item{relax_priority}{Relaxation priority vector}
}
\description{
Displays general information at the start of filtering.
}
\keyword{internal}
