% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_modwt.R
\name{modwt_reconstruct}
\alias{modwt_reconstruct}
\title{Inverse MODWT}
\usage{
modwt_reconstruct(wt)
}
\arguments{
\item{wt}{A list of class \code{"modwt_result"} from \code{\link{modwt_decompose}},
or a raw waveslim modwt object.}
}
\value{
Numeric vector. Reconstructed time series.
}
\description{
Reconstructs a time series from its MODWT coefficients.
}
\keyword{internal}
