% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warm_filtering.R
\name{pad_right}
\alias{pad_right}
\title{Pad string to the right}
\usage{
pad_right(x, width)
}
\arguments{
\item{x}{Character string}

\item{width}{Target width}
}
\value{
Padded character string
}
\description{
Pad string to the right
}
\keyword{internal}
