% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{pick_year_window}
\alias{pick_year_window}
\title{Pick a random contiguous window of years of length n_years from a year vector}
\usage{
pick_year_window(years, n_years)
}
\description{
Pick a random contiguous window of years of length n_years from a year vector
}
\keyword{internal}
