% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{plot.prcp_qm_diagnostics}
\alias{plot.prcp_qm_diagnostics}
\title{Plot diagnostics for precipitation QM}
\usage{
\method{plot}{prcp_qm_diagnostics}(x, which = c("all", "moments", "quantiles", "extremes", "monthly"), ...)
}
\description{
Plot diagnostics for precipitation QM
}
