% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare data for diagnostic plotting (obs vs sim)}
\usage{
prepare_plot_data(obs_results, sim_results, vars)
}
\description{
Prepare data for diagnostic plotting (obs vs sim)
}
\keyword{internal}
