% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_modwt.R
\name{print.wavelet_additive}
\alias{print.wavelet_additive}
\title{Print Method for Additive Wavelet Analysis}
\usage{
\method{print}{wavelet_additive}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"wavelet_additive"}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns x.
}
\description{
Print Method for Additive Wavelet Analysis
}
