% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{process_observed_data}
\alias{process_observed_data}
\title{Process observed weather data}
\usage{
process_observed_data(
  daily.obs,
  variables,
  n_grids,
  wet.quantile,
  extreme.quantile
)
}
\description{
Process observed weather data
}
\keyword{internal}
