% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileMapping.R
\name{quantileMapping}
\alias{quantileMapping}
\title{Function for perturbing climate statistics using Quantile-Mapping}
\usage{
quantileMapping(
  value = NULL,
  date = NULL,
  mean.change = NULL,
  var.change = NULL,
  mon.ts = NULL,
  year.ts = NULL,
  step.change = TRUE,
  reltol = 1e-05
)
}
\arguments{
\item{value}{To be completed...}

\item{date}{To be completed...}

\item{mean.change}{placeholder}

\item{var.change}{placeholder}

\item{mon.ts}{placeholder}

\item{year.ts}{placeholder}

\item{step.change}{placeholder}

\item{reltol}{placeholder}
}
\value{

}
\description{
Function for perturbing climate statistics using Quantile-Mapping
}
