% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileMappingDailyTS.R
\name{quantileMappingDailyTS}
\alias{quantileMappingDailyTS}
\title{Function for perturbing climate statistics using Quantile-Mapping}
\usage{
quantileMappingDailyTS(
  parms = NULL,
  base.ts = NULL,
  base.dist = "gamma",
  change.operator = "multiply"
)
}
\arguments{
\item{parms}{List of parameters to be passed.}

\item{base.ts}{Base time-series to be used.}

\item{base.dist}{Base distribution selected.}

\item{change.operator}{A character string to specify the type of perturb change operator.}
}
\value{

}
\description{
Function for perturbing climate statistics using Quantile-Mapping
}
