% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quiet}
\alias{quiet}
\title{A function to surppresses the output messages}
\usage{
quiet(x)
}
\arguments{
\item{x}{An R function to surpress the output messages}
}
\value{

}
\description{
Function first suggested by Hadley Wickham http://r.789695.n4.nabble.com/Suppressing-output-e-g-from-cat-td859876.html
}
