% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFromNetcdf.R
\name{readFromNetcdf}
\alias{readFromNetcdf}
\title{Function to read climate data from netcdf files}
\usage{
readFromNetcdf(
  in.path = NULL,
  in.file = NULL,
  dim.names = NULL,
  variables = NULL,
  origin.date = NULL
)
}
\arguments{
\item{in.path}{A character string to define the path of template netcdf file.}

\item{in.file}{A character string to define the name of the template netcdf file.}

\item{dim.names}{A list object specifying the dimension names.}

\item{variables}{A list object specifying variable names}

\item{origin.date}{A character string to specify the date of origin}
}
\value{

}
\description{
Function to read climate data from netcdf files
}
