% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNetcdf.R
\name{readNetcdf}
\alias{readNetcdf}
\title{Function to read climate data from netcdf files}
\usage{
readNetcdf(
  nc.path = NULL,
  nc.file = NULL,
  nc.dimnames = NULL,
  nc.variables = NULL,
  origin.date = NULL,
  leap.year = TRUE,
  omit.nans = TRUE
)
}
\arguments{
\item{nc.path}{A character string to define the path of template netcdf file.}

\item{nc.file}{A character string to define the name of the template netcdf file.}

\item{nc.dimnames}{A list object specifying the dimension names.}

\item{origin.date}{A character string to specify the date of origin}

\item{leap.year}{placeholder}

\item{variables}{A list object specifying variable names}
}
\value{

}
\description{
Function to read climate data from netcdf files
}
