% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleDailyWeather.R
\name{resampleDailyWeather}
\alias{resampleDailyWeather}
\title{Dissaggregation function}
\usage{
resampleDailyWeather(
  k1 = NULL,
  ymax = NULL,
  PRCP_FINAL_ANNUAL_SIM = NULL,
  ANNUAL_PRCP = NULL,
  WATER_YEAR_A,
  WATER_YEAR_D,
  PRCP,
  TEMP,
  DATE_D,
  MONTH_D,
  YEAR_D,
  MONTH_DAY_D,
  month_list,
  water_year_start,
  water_year_end,
  y_sample_size = 20,
  SIM_LENGTH = NULL,
  kk = NULL,
  thresh1 = 0.3,
  extreme_quantile = 0.8,
  MONTH_SIM = NULL,
  WATER_YEAR_SIM = NULL,
  START_YEAR_SIM = NULL,
  DAY_SIM = NULL
)
}
\arguments{
\item{k1}{placeholder}

\item{ymax}{placeholder}

\item{PRCP_FINAL_ANNUAL_SIM}{placeholder}

\item{ANNUAL_PRCP}{placeholder}

\item{WATER_YEAR_A}{placeholder}

\item{WATER_YEAR_D}{placeholder}

\item{PRCP}{placeholder}

\item{TEMP}{placeholder}

\item{DATE_D}{placeholder}

\item{MONTH_D}{placeholder}

\item{YEAR_D}{placeholder}

\item{MONTH_DAY_D}{placeholder}

\item{month_list}{placeholder}

\item{water_year_start}{placeholder}

\item{water_year_end}{placeholder}

\item{y_sample_size}{placeholder}

\item{SIM_LENGTH}{placeholder}
}
\value{

}
\description{
Dissaggregation function
}
