% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleDates_OLD.R
\name{resampleDates_OLD}
\alias{resampleDates_OLD}
\title{Dissaggregation function}
\usage{
resampleDates_OLD(
  PRCP_FINAL_ANNUAL_SIM = NULL,
  ANNUAL_PRCP = NULL,
  PRCP = NULL,
  TEMP = NULL,
  TMAX = NULL,
  TMIN = NULL,
  START_YEAR_SIM = NULL,
  k1 = NULL,
  ymax = NULL,
  dates.d = NULL,
  sim.dates.d = NULL,
  YEAR_D = NULL,
  month.start = NULL,
  knn.annual.sample.num = 50,
  wet.quantile = 0.2,
  extreme.quantile = 0.8,
  dry.spell.change = rep(1, 12),
  wet.spell.change = rep(1, 12),
  seed = NULL
)
}
\arguments{
\item{PRCP_FINAL_ANNUAL_SIM}{placeholder}

\item{ANNUAL_PRCP}{placeholder}

\item{PRCP}{placeholder}

\item{TEMP}{placeholder}

\item{TMAX}{placeholder}

\item{TMIN}{placeholder}

\item{START_YEAR_SIM}{placeholder}

\item{k1}{placeholder}

\item{ymax}{placeholder}

\item{dates.d}{placeholder}

\item{sim.dates.d}{placeholder}

\item{YEAR_D}{placeholder}

\item{month.start}{placeholder}

\item{knn.annual.sample.num}{placeholder}

\item{wet.quantile}{Precipitation threshold (mm) to distinguish between wet and dry days}

\item{extreme.quantile}{quantile threshold to distinguish between extreme wet days}

\item{dry.spell.change}{placeholder}

\item{wet.spell.change}{placeholder}

\item{seed}{random seed value}
}
\value{
A vector of `Date` objects representing the resampled daily dates for the simulation period.
}
\description{
Dissaggregation function
}
