% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGriddedWeather.R
\name{simulateGriddedWeather}
\alias{simulateGriddedWeather}
\title{Main wrapper function for daily weather generation}
\usage{
simulateGriddedWeather(
  proj.name = NULL,
  output.dir = NULL,
  wg.date.begin = NULL,
  sim.date.begin = NULL,
  wg.vars = NULL,
  wg.var.labs = NULL,
  wg.var.units = NULL,
  warm.var = NULL,
  ssig = 0.9,
  ymax = 40,
  rmax = 2000,
  nmax = 5,
  PARCC = NULL,
  nc.path = NULL,
  nc.file = NULL,
  nc.dimnames = NULL,
  perturb.file = NULL,
  validate = FALSE
)
}
\arguments{
\item{proj.name}{Placeholder.}

\item{wg.date.begin}{Placeholder.}

\item{sim.date.begin}{Placeholder.}

\item{wg.vars}{Placeholder.}

\item{wg.var.labs}{Placeholder.}

\item{wg.var.units}{Placeholder.}

\item{warm.var}{Placeholder.}

\item{ssig}{Placeholder.}

\item{ymax}{Placeholder.}

\item{rmax}{Placeholder.}

\item{nmax}{Placeholder.}

\item{PARCC}{Placeholder.}

\item{nc.path}{Placeholder.}

\item{nc.file}{Placeholder.}

\item{nc.dimnames}{Placeholder.}

\item{validate}{Placeholder.}

\item{out_path}{Placeholder.}

\item{wg.date.end}{Placeholder.}
}
\value{

}
\description{
Main wrapper function for daily weather generation
}
