% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateWeather.R
\name{simulateWeather}
\alias{simulateWeather}
\title{Simulate gridded weather function}
\usage{
simulateWeather(
  climate.data = NULL,
  climate.grid = NULL,
  year.start = NULL,
  year.num = NULL,
  month.start = 1,
  variable.names = NULL,
  variable.labels = NULL,
  variable.units = NULL,
  sim.year.start = 2020,
  sim.year.num = 40,
  realization.num = 5,
  warm.variable = "precip",
  warm.signif.level = 0.9,
  warm.sample.size = 10000,
  save.warm.results = TRUE,
  knn.annual.sample.size = 50,
  evaluate.model = FALSE,
  evaluate.grid.num = 20,
  apply.delta.changes = TRUE,
  apply.step.changes = TRUE,
  delta.precip = NULL,
  delta.temp = NULL,
  save.scenario.matrix = TRUE,
  output.path = NULL,
  output.ncfile.template = NULL,
  output.ncfile.prefix = "clim_change_rlz",
  ...
)
}
\arguments{
\item{climate.data}{placeholder}

\item{climate.grid}{placeholder}

\item{year.start}{placeholder}

\item{year.num}{placeholder}

\item{month.start}{placeholder}

\item{variable.names}{placeholder}

\item{variable.labels}{placeholder}

\item{variable.units}{placeholder}

\item{sim.year.start}{placeholder}

\item{sim.year.num}{placeholder}

\item{realization.num}{placeholder}

\item{warm.variable}{placeholder}

\item{warm.signif.level}{placeholder}

\item{warm.sample.size}{placeholder}

\item{save.warm.results}{placeholder}

\item{knn.annual.sample.size}{placeholder}

\item{evaluate.model}{placeholder}

\item{evaluate.grid.num}{placeholder}

\item{apply.delta.changes}{placeholder}

\item{apply.step.changes}{placeholder}

\item{delta.precip}{placeholder}

\item{delta.temp}{placeholder}

\item{save.scenario.matrix}{placeholder}

\item{output.path}{placeholder}

\item{output.ncfile.template}{placeholder}

\item{output.ncfile.prefix}{placeholder}

\item{...}{placeholder}
}
\value{

}
\description{
Simulate gridded weather function
}
