% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{skewness}
\alias{skewness}
\title{Skewness function (from package e1071)}
\usage{
skewness(x, na.rm = FALSE, type = 3)
}
\arguments{
\item{x}{a numeric vector containing the values whose skewness is to be computed.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{type}{an integer between 1 and 3 selecting one of the algorithms for computing skewness detailed below.}
}
\value{

}
\description{
Skewness function (from package e1071)
}
