% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{standardize_obs_sim_periods}
\alias{standardize_obs_sim_periods}
\title{Standardize obs and sim to same full-year length via random windowing}
\usage{
standardize_obs_sim_periods(obs_data, sim_data, n_realizations)
}
\description{
Standardize obs and sim to same full-year length via random windowing
}
\keyword{internal}
