% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{summary.prcp_qm_diagnostics}
\alias{summary.prcp_qm_diagnostics}
\title{Summarize precipitation QM diagnostics}
\usage{
\method{summary}{prcp_qm_diagnostics}(object, ...)
}
\arguments{
\item{object}{A \code{"prcp_qm_diagnostics"} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Prints a text summary for a \code{"prcp_qm_diagnostics"} object.
}
