% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate inputs for weather assessment}
\usage{
validate_inputs(
  daily.sim,
  daily.obs,
  variables,
  realization.num,
  wet.quantile,
  extreme.quantile
)
}
\description{
Validate inputs for weather assessment
}
\keyword{internal}
