% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{validate_weather_generator_inputs}
\alias{validate_weather_generator_inputs}
\title{Validate inputs for evaluate_weather_generator()}
\usage{
validate_weather_generator_inputs(
  sim_data,
  obs_data,
  vars,
  n_realizations,
  wet_q,
  extreme_q,
  max_grid,
  seed
)
}
\description{
Validate inputs for evaluate_weather_generator()
}
\keyword{internal}
