% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletAnalysis.R
\name{waveletAnalysis}
\alias{waveletAnalysis}
\title{Wavelet Analysis Function}
\usage{
waveletAnalysis(
  variable = NULL,
  variable.unit = "mm",
  signif.level = 0.9,
  noise = "white",
  plot = FALSE,
  out.path = NULL
)
}
\arguments{
\item{variable}{A numeric vector of time-series of variables, for example, time-series of annual precipitation.}

\item{variable.unit}{A character string to define the unit of the variable.}

\item{signif.level}{A numeric value to set the siginificance level of the wavelet analysis (Default= 0.90).}

\item{noise}{A character string defining the type of background noise. Currently either "white" (default) or "red".}

\item{plot}{Draw plot}

\item{out.path}{Output path}
}
\value{

}
\description{
\code{waveletAnalysis} returns the wavelet analysis results.
}
\details{
To be completed later...
}
