% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletAR.R
\name{waveletAR}
\alias{waveletAR}
\title{A function for Wavelet Autoregressive Modeling (WARM)}
\usage{
waveletAR(wavelet.comps = NULL, num.years = NULL, num.realizations = 1000)
}
\arguments{
\item{wavelet.comps}{A list object, with different components corresponding to low-frequency signals and the noise}

\item{num.years}{A numeric value defining the desired length of simulated annual time-series}

\item{num.realizations}{A numeric value defining the number of synthetic series to be produced.}
}
\description{
A function for Wavelet Autoregressive Modeling (WARM)
}
