% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletDecomposition.R
\name{waveletDecomposition}
\alias{waveletDecomposition}
\title{Function for weather generator decomposition}
\usage{
waveletDecomposition(
  variable = NULL,
  signif.period.list = NULL,
  background.noise = "white",
  signif.level = 0.8
)
}
\arguments{
\item{variable}{A vector of time-series of weather variables.}

\item{signif.period.list}{Significant low-frequency periods in the original time-series.}

\item{background.noise}{A logical specifying the type of background noise.}

\item{signif.level}{Significance level for the wavelet analysis.}
}
\value{

}
\description{
Function for weather generator decomposition
}
