% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletGlobalSpectrumPlot.R
\name{waveletGlobalSpectrumPlot}
\alias{waveletGlobalSpectrumPlot}
\title{ggplot wrapper for wavelet spectra plot}
\usage{
waveletGlobalSpectrumPlot(
  power.period,
  power.signif,
  power.obs,
  power.sim = NULL
)
}
\arguments{
\item{power.period}{A numeric vector of fourier periods.}

\item{power.signif}{A numeric value to set the sigicance level.}

\item{power.obs}{A numeric vector of observed global wavelet spectra.}

\item{power.sim}{A numeric matrix of simulated global wavelet spectrum.}
}
\value{
A ggplot2 object
}
\description{
This function is simply a ggplot wrapper for displaying results from a wavelet analysis.
}
\details{
place holder
}
