% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletPlot.R
\name{waveletPlot}
\alias{waveletPlot}
\title{A ggplot wrapper for wavelet spectra plot}
\usage{
waveletPlot(
  power.period = NULL,
  power.signif = NULL,
  power.obs = NULL,
  power.sim = NULL
)
}
\arguments{
\item{power.period}{A numeric vector of fourier periods from wavelet analysis}

\item{power.signif}{A numeric value to set the sigicance level of the spectral analysis}

\item{power.obs}{A numeric vector of observed  global wavelet spectra.}

\item{power.sim}{A numeric matrix of simulated global wavelet spectrum. Each column is an independent observation}
}
\value{
A ggplot2 object
}
\description{
ggplot wrapper for visualization of wavelet spectral analysis
}
\details{
place holder
}
