% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToNetcdf.R
\name{writeToNetcdf}
\alias{writeToNetcdf}
\title{A function to write a netcdf file from tidydata}
\usage{
writeToNetcdf(
  in.path = NULL,
  in.file = NULL,
  out.path = NULL,
  out.file = NULL,
  out.variables = NULL,
  dim.names = NULL,
  origin.date = NULL,
  gridded.data = NULL,
  verbose = FALSE
)
}
\arguments{
\item{in.path}{A character string to define the path of template netcdf file.}

\item{in.file}{A character string to define the name of the template netcdf file.}

\item{out.path}{A character string to define the path of resulting netcdf file.}

\item{out.file}{A character string to define the name of resulting netcdf file.}

\item{out.variables}{A character vector to define which variables to be outputted to the netcdf file.}

\item{dim.names}{A character vector of dimension names.}

\item{origin.date}{A character string to set the origin date in the resulting netcdf file.}

\item{gridded.data}{A multidimensional array to be written to the netcdf file.}

\item{verbose}{A logical to print more information on the screen.}
}
\value{

}
\description{
A function to write a netcdf file from tidydata
}
