% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageSpellLength.R
\name{averageSpellLength}
\alias{averageSpellLength}
\title{Calculate Average Spell Length}
\usage{
averageSpellLength(x, threshold = 0, below = TRUE)
}
\arguments{
\item{x}{A numeric vector representing daily weather values (e.g., precipitation).}

\item{threshold}{A numeric threshold used to define spell classification.}

\item{below}{Logical. If TRUE, calculates dry spells (values less than threshold). If FALSE, calculates wet spells (values more than threshold).}
}
\value{
A single numeric value representing the average spell length.
}
\description{
Computes the average length of either wet or dry spells based on a numeric input vector and a threshold.
}
\keyword{internal}
