% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSpellLength.R
\name{averageSpellLength}
\alias{averageSpellLength}
\title{Calculate Spell Length}
\usage{
averageSpellLength(x, threshold = 0, below = TRUE)
}
\arguments{
\item{x}{placeholder}

\item{threshold}{placeholder}

\item{below}{placeholder}
}
\description{
Calculate Spell Length
}
