% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateSurface.R
\name{climateSurface}
\alias{climateSurface}
\title{Climate Response Surface Generator}
\usage{
climateSurface(
  str.data = NULL,
  gcm.data = NULL,
  variable.x = NULL,
  variable.y = NULL,
  variable.z = NULL,
  threshold.z = NULL,
  plot.title = "change in variable",
  variable.x.label = expression(Delta ~ "Precipitation"),
  variable.y.label = expression(Delta ~ "Temperature"),
  failure.direction = 1,
  gcm.scenario.list = c("rcp26", "rcp45", "rcp60", "rcp85"),
  gcm.bvnorm.levels = FALSE,
  gcm.transparency = 0.75,
  gcm.legend = TRUE,
  variable.z.breaks = NULL,
  variable.x.breaks = NULL,
  variable.y.breaks = NULL,
  contour.num = 15,
  text.scale = 0.6,
  multi.panel = FALSE,
  panel.variable = NULL,
  panel.variable.levels = NULL
)
}
\arguments{
\item{str.data}{placeholder}

\item{gcm.data}{placeholder}

\item{variable.x}{placeholder}

\item{variable.y}{placeholder}

\item{variable.z}{placeholder}

\item{threshold.z}{placeholder}

\item{plot.title}{placeholder}

\item{variable.x.label}{placeholder}

\item{variable.y.label}{placeholder}

\item{failure.direction}{placeholder}

\item{gcm.scenario.list}{placeholder}

\item{gcm.bvnorm.levels}{placeholder}

\item{gcm.transparency}{placeholder}

\item{gcm.legend}{placeholder}

\item{variable.z.breaks}{placeholder}

\item{variable.x.breaks}{placeholder}

\item{variable.y.breaks}{placeholder}

\item{contour.num}{placeholder}

\item{text.scale}{placeholder}

\item{multi.panel}{placeholder}

\item{panel.variable}{placeholder}

\item{panel.variable.levels}{placeholder}

\item{variable.z.min}{placeholder}

\item{variable.z.max}{placeholder}
}
\description{
Climate Response Surface Generator
}
