% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateSurface.R
\name{climateSurface}
\alias{climateSurface}
\title{Climate Response Surface Generator}
\usage{
climateSurface(
  str.data = NULL,
  gcm.data = NULL,
  variable.x = NULL,
  variable.y = NULL,
  variable.z = NULL,
  threshold.z = NULL,
  plot.title = "change in variable",
  variable.x.label = "change in temperature",
  variable.y.label = "change in precipitation",
  failure.direction = 1,
  gcm.scenario.list = c("rcp26", "rcp45", "rcp60", "rcp85"),
  gcm.bivariate.dist = FALSE,
  gcm.transparency = 0.75,
  gcm.legend = TRUE,
  variable.z.min = NULL,
  variable.z.max = NULL,
  variable.z.bin = 15,
  variable.z.min.legend = NULL,
  variable.z.max.legend = NULL,
  variable.z.bin.legend = 11,
  variable.x.breaks = NULL,
  variable.y.breaks = NULL,
  text.scale = 0.8
)
}
\arguments{
\item{str.data}{placeholder}

\item{gcm.data}{placeholder}

\item{variable.x}{placeholder}

\item{variable.y}{placeholder}

\item{variable.z}{placeholder}

\item{threshold.z}{placeholder}

\item{plot.title}{placeholder}

\item{variable.x.label}{placeholder}

\item{variable.y.label}{placeholder}

\item{failure.direction}{placeholder}

\item{gcm.scenario.list}{placeholder}

\item{gcm.bivariate.dist}{placeholder}

\item{gcm.transparency}{placeholder}

\item{gcm.legend}{placeholder}

\item{variable.z.min}{placeholder}

\item{variable.z.max}{placeholder}

\item{variable.z.bin}{placeholder}

\item{variable.z.min.legend}{placeholder}

\item{variable.z.max.legend}{placeholder}

\item{variable.z.bin.legend}{placeholder}

\item{variable.x.breaks}{placeholder}

\item{variable.y.breaks}{placeholder}

\item{text.scale}{placeholder}
}
\description{
Climate Response Surface Generator
}
