% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateSurfaceBatch.R
\name{climateSurfaceBatch}
\alias{climateSurfaceBatch}
\title{Batch create climate response surfaces from CST Toolbox applications}
\usage{
climateSurfaceBatch(
  save.dir = NULL,
  str.data = NULL,
  gcm.data = NULL,
  gcm.legend = TRUE,
  gcm.future.period = "near",
  metrics = NULL,
  locations = NULL,
  location.labels = NULL,
  metric.labels = NULL,
  metric.direction = NULL,
  relative.results = TRUE,
  ...
)
}
\arguments{
\item{save.dir}{placeholder}

\item{str.data}{placeholder}

\item{gcm.data}{placeholder}

\item{gcm.future.period}{placeholder}

\item{metrics}{placeholder}

\item{metric.labels}{placeholder}

\item{metric.direction}{placeholder}

\item{...}{}
}
\description{
Batch create climate response surfaces from CST Toolbox applications
}
