% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{compute_correlation_matrix_anom}
\alias{compute_correlation_matrix_anom}
\title{Compute correlation matrix on monthly anomalies}
\usage{
compute_correlation_matrix_anom(data, variables)
}
\description{
Compute correlation matrix on monthly anomalies
}
\keyword{internal}
