% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quantile_mapping.R
\name{compute_dryday_diagnostics}
\alias{compute_dryday_diagnostics}
\title{Compute dry day diagnostics}
\usage{
compute_dryday_diagnostics(value.original, value.adjusted, threshold)
}
\description{
Compute dry day diagnostics
}
\keyword{internal}
