% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_diagnostics.R
\name{compute_dryday_diagnostics}
\alias{compute_dryday_diagnostics}
\title{Compute dry-day and wet-day frequency diagnostics (intended preserved)}
\usage{
compute_dryday_diagnostics(prcp_ref, prcp_adj, wet_thresh, tol_freq_abs = 0.01)
}
\description{
Compute dry-day and wet-day frequency diagnostics (intended preserved)
}
\keyword{internal}
