% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{compute_dryday_diagnostics}
\alias{compute_dryday_diagnostics}
\title{Compute dry-day and wet-day frequency diagnostics}
\usage{
compute_dryday_diagnostics(prcp_ref, prcp_adj, wet_thresh)
}
\description{
Compute dry-day and wet-day frequency diagnostics
}
\keyword{internal}
