% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quantile_mapping.R
\name{compute_extreme_diagnostics}
\alias{compute_extreme_diagnostics}
\title{Compute extreme value diagnostics}
\usage{
compute_extreme_diagnostics(value.original, value.adjusted)
}
\description{
Compute extreme value diagnostics
}
\keyword{internal}
