% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_diagnostics.R
\name{compute_extreme_diagnostics}
\alias{compute_extreme_diagnostics}
\title{Compute extreme-value diagnostics (tail ratios)}
\usage{
compute_extreme_diagnostics(prcp_ref, prcp_adj)
}
\description{
Compute extreme-value diagnostics (tail ratios)
}
\keyword{internal}
