% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{compute_extreme_diagnostics}
\alias{compute_extreme_diagnostics}
\title{Compute extreme-value diagnostics for quantile-mapped precipitation}
\usage{
compute_extreme_diagnostics(prcp_ref, prcp_adj)
}
\description{
Compute extreme-value diagnostics for quantile-mapped precipitation
}
\keyword{internal}
