% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{compute_grouped_statistics}
\alias{compute_grouped_statistics}
\title{Compute grouped statistics (long-form output)}
\usage{
compute_grouped_statistics(df, vars, group_var, stat_funs)
}
\description{
Compute grouped statistics (long-form output)
}
\keyword{internal}
