% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quantile_mapping.R
\name{compute_moment_diagnostics}
\alias{compute_moment_diagnostics}
\title{Compute moment diagnostics}
\usage{
compute_moment_diagnostics(value.original, value.adjusted, nonzero.mask)
}
\description{
Compute moment diagnostics
}
\keyword{internal}
