% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{compute_moment_diagnostics}
\alias{compute_moment_diagnostics}
\title{Compute moment diagnostics for quantile-mapped precipitation}
\usage{
compute_moment_diagnostics(prcp_ref, prcp_adj, mask)
}
\description{
Compute moment diagnostics for quantile-mapped precipitation
}
\keyword{internal}
