% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quantile_mapping.R
\name{compute_monthly_diagnostics}
\alias{compute_monthly_diagnostics}
\title{Compute monthly diagnostics}
\usage{
compute_monthly_diagnostics(
  value.original,
  value.adjusted,
  mon.ts,
  mean.change = NULL,
  var.change = NULL,
  year.ts = NULL
)
}
\description{
Compute monthly diagnostics
}
\keyword{internal}
