% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{compute_quantile_diagnostics}
\alias{compute_quantile_diagnostics}
\title{Compute quantile diagnostics for quantile-mapped precipitation}
\usage{
compute_quantile_diagnostics(prcp_ref, prcp_adj, probs)
}
\description{
Compute quantile diagnostics for quantile-mapped precipitation
}
\keyword{internal}
