% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_diagnostics.R
\name{compute_spell_diagnostics}
\alias{compute_spell_diagnostics}
\title{Compute wet and dry spell-length diagnostics (intended preserved)}
\usage{
compute_spell_diagnostics(prcp_ref, prcp_adj, wet_thresh, tol_mean_ratio = 0.2)
}
\description{
Compute wet and dry spell-length diagnostics (intended preserved)
}
\keyword{internal}
