% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{compute_spell_diagnostics}
\alias{compute_spell_diagnostics}
\title{Compute wet and dry spell-length diagnostics}
\usage{
compute_spell_diagnostics(prcp_ref, prcp_adj, wet_thresh)
}
\description{
Compute wet and dry spell-length diagnostics
}
\keyword{internal}
