% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quantile_mapping.R
\name{compute_spell_lengths}
\alias{compute_spell_lengths}
\title{Compute spell lengths}
\usage{
compute_spell_lengths(x, threshold, below = TRUE)
}
\description{
Compute spell lengths
}
\keyword{internal}
