% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quantile_mapping.R
\name{compute_summary_metrics}
\alias{compute_summary_metrics}
\title{Compute overall summary metrics}
\usage{
compute_summary_metrics(
  moments,
  quantile.metrics,
  extreme.metrics,
  spell.metrics,
  dryday.metrics
)
}
\description{
Compute overall summary metrics
}
\keyword{internal}
