% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_diagnostics.R
\name{compute_summary_metrics}
\alias{compute_summary_metrics}
\title{Compute summary metrics (no qualitative grades; no scoring)}
\usage{
compute_summary_metrics(
  moments,
  quant_metrics,
  extreme_metrics,
  spell_metrics,
  dryday_metrics
)
}
\description{
Compute summary metrics (no qualitative grades; no scoring)
}
\keyword{internal}
