% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_mapping.R
\name{compute_target_parameters}
\alias{compute_target_parameters}
\title{Compute target distribution parameters from base params and change factors}
\usage{
compute_target_parameters(
  base_gamma,
  mean_factor,
  var_factor,
  months_ok,
  n_years
)
}
\description{
Compute target distribution parameters from base params and change factors
}
\keyword{internal}
