% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{compute_temporal_diagnostics}
\alias{compute_temporal_diagnostics}
\title{Compute temporal diagnostics for quantile-mapped precipitation}
\usage{
compute_temporal_diagnostics(prcp_ref, prcp_adj, month, year, mask)
}
\description{
Compute temporal diagnostics for quantile-mapped precipitation
}
\keyword{internal}
