% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{compute_timeseries_statistics}
\alias{compute_timeseries_statistics}
\title{Compute time series statistics}
\usage{
compute_timeseries_statistics(data, variables, mc.thresholds)
}
\description{
Compute time series statistics
}
\keyword{internal}
