% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{create_annual_precip_plot}
\alias{create_annual_precip_plot}
\title{Create annual precipitation plot}
\usage{
create_annual_precip_plot(
  stats.annual.aavg.sim,
  stats.annual.aavg.obs,
  plot.config,
  show.title,
  save.plots,
  output.path
)
}
\description{
Create annual precipitation plot
}
\keyword{internal}
