% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warm_filtering.R
\name{criteria_string_compact}
\alias{criteria_string_compact}
\title{Compact criteria string for a filter}
\usage{
criteria_string_compact(
  filter_name,
  bounds,
  tail_metrics,
  wavelet_active,
  wavelet_pars
)
}
\description{
Compact criteria string for a filter
}
\keyword{internal}
