% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_warm.R
\name{.fast_col_sd}
\alias{.fast_col_sd}
\title{Fast column standard deviations}
\usage{
.fast_col_sd(x)
}
\arguments{
\item{x}{Numeric matrix.}
}
\value{
Numeric vector of column standard deviations.
}
\description{
Fast column standard deviations
}
\keyword{internal}
