% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{.validate_assessment_inputs}
\alias{.validate_assessment_inputs}
\title{Validate inputs for weather assessment}
\usage{
.validate_assessment_inputs(
  daily_sim,
  daily_obs,
  vars,
  n_realizations,
  wet_q,
  extreme_q,
  verbose,
  parallel,
  n_cores
)
}
\description{
Validate inputs for weather assessment
}
\keyword{internal}
