% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateWegen.R
\name{evaluateWegen}
\alias{evaluateWegen}
\title{Compare observed and simulated climate statistics}
\usage{
evaluateWegen(
  daily.sim = NULL,
  daily.obs = NULL,
  output.path = NULL,
  variables = NULL,
  variable.labels = NULL,
  variable.units = NULL,
  realization.num = NULL,
  wet.quantile = 0.2,
  extreme.quantile = 0.8,
  show.title = TRUE,
  save.plots = TRUE
)
}
\arguments{
\item{daily.sim}{A matrix of daily time-series of weather variables}

\item{daily.obs}{A vector of daily time-series of observed weather series}

\item{output.path}{A character string to define the path of resulting netcdf file}

\item{variables}{A vector object specifying the names of weather variables}

\item{variable.labels}{A vector object specifying the names of weather variable labels}

\item{variable.units}{A vector object specifying the names of weather variable units}

\item{realization.num}{number of weather realizations}

\item{wet.quantile}{precip. quantile for calculating monthly wet state threshold}

\item{extreme.quantile}{precip. quantile for calculating monthly extremely wet state threshold}

\item{show.title}{PLACEHOLDER}

\item{save.plots}{PLACEHOLDER}
}
\value{
returns a list of ggplot2 objects and figures saved to the desired output path
}
\description{
A short description...
}
\details{
Additional details...
}
