% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{filter_correlations}
\alias{filter_correlations}
\title{Filter correlations by type (cross-grid or inter-variable)}
\usage{
filter_correlations(cor_data, same_var, same_id, allowed_pairs, pair_type)
}
\description{
Filter correlations by type (cross-grid or inter-variable)
}
\keyword{internal}
