% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb_prcp_qm.R
\name{fit_monthly_distributions}
\alias{fit_monthly_distributions}
\title{Fit gamma distributions for each month}
\usage{
fit_monthly_distributions(prcp_wet, month_wet, months_ok, fit_method, verbose)
}
\description{
Fit gamma distributions for each month
}
\keyword{internal}
