% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_mapping.R
\name{fit_monthly_distributions}
\alias{fit_monthly_distributions}
\title{Fit gamma distributions for each month}
\usage{
fit_monthly_distributions(value.nz, mon.nz, valid.months, fit.method, verbose)
}
\description{
Fit gamma distributions for each month
}
\keyword{internal}
