% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{generate_symmetric_dummy_points}
\alias{generate_symmetric_dummy_points}
\title{Generate symmetric dummy points for equal axis scaling}
\usage{
generate_symmetric_dummy_points(
  df,
  facet.var = "variable",
  x.col = "Observed",
  y.col = "Simulated"
)
}
\description{
Generate symmetric dummy points for equal axis scaling
}
\keyword{internal}
