% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateWeatherSeries.R
\name{generateWeatherSeries}
\alias{generateWeatherSeries}
\title{Simulate gridded weather function}
\usage{
generateWeatherSeries(
  weather.data = NULL,
  weather.grid = NULL,
  weather.date = NULL,
  variable.names = NULL,
  variable.labels = NULL,
  variable.units = NULL,
  sim.year.num = NULL,
  sim.year.start = 2020,
  month.start = 1,
  realization.num = 5,
  warm.variable = "precip",
  warm.signif.level = 0.9,
  warm.sample.num = 5000,
  warm.subset.criteria = NULL,
  knn.sample.num = 100,
  mc.wet.threshold = 0.3,
  mc.extreme.quantile = 0.8,
  evaluate.model = FALSE,
  evaluate.grid.num = 20,
  output.path = getwd(),
  seed = NULL
)
}
\arguments{
\item{weather.data}{list of data frames of daily weather observations per grid cell. Each data frame, columns are weather variables and rows are daily values.}

\item{weather.grid}{Data frame of grid cells. Each grid cell is assigned an id starting from 1, x and y coordinate index value, and x and y coordinates.}

\item{weather.date}{placeholder}

\item{variable.names}{vector of names for the variables to be included in the weather generator}

\item{variable.labels}{vector of labels for the weather variables (optional). If no values provided, it is labels will be same as the names}

\item{variable.units}{vector of units for each of the weather variables (optional). If no values provided, a blank vector is used.}

\item{sim.year.num}{numeric value indicating the desired total number of years of simulated weather realizations}

\item{sim.year.start}{numeric value indicating the starting year of the generated time-series}

\item{month.start}{the first month of the year (default value is 1). Use a value other than 1 for water-year based analyses}

\item{realization.num}{number of natural variability realizations to be generated.}

\item{warm.variable}{the name of the variable for the wavelet autoregressive mode. Default value is precipitation variable.}

\item{warm.signif.level}{the significance level for the warm model.}

\item{warm.sample.num}{number of annual sequeces to be generated from the the warm model}

\item{warm.subset.criteria}{placeholder}

\item{knn.sample.num}{number of knn years to be sampled}

\item{mc.wet.threshold}{placeholder}

\item{mc.extreme.quantile}{placeholder}

\item{evaluate.model}{logical value indicating wether to save model evaluation plots}

\item{evaluate.grid.num}{Number of grid cells to be sampled in the evaluation plots}

\item{output.path}{output path for the weather generator results (string)}

\item{seed}{placeholder}
}
\value{

}
\description{
Description goes here....
}
