% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPI.R
\name{getPI}
\alias{getPI}
\title{Calculate markov chain probobabilities}
\usage{
getPI(p00, p01, p02, p10, p11, p12, p20, p21, p22)
}
\arguments{
\item{p00}{placeholder}

\item{p01}{placeholder}

\item{p02}{placeholder}

\item{p10}{placeholder}

\item{p11}{placeholder}

\item{p12}{placeholder}

\item{p20}{placeholder}

\item{p21}{placeholder}

\item{p22}{placeholder}
}
\description{
Calculate markov chain probobabilities
}
