% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hargreavesPET.R
\name{hargreavesPet}
\alias{hargreavesPet}
\title{Potential evapotranspiration (PET) by hargreaves method}
\usage{
hargreavesPet(months, temp, tdiff, lat)
}
\arguments{
\item{months}{To be completed...}

\item{temp}{To be completed...}

\item{tdiff}{To be completed...}

\item{lat}{To be completed...}
}
\value{
the output is a vector of PET values
}
\description{
\url{http://www.civil.uwaterloo.ca/watflood/manual/02_03_2.htm}
}
