% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeClimateChanges.R
\name{imposeClimateChanges}
\alias{imposeClimateChanges}
\title{Perturb weather realizations to reflect climate change}
\usage{
imposeClimateChanges(
  climate.data = NULL,
  climate.grid = NULL,
  sim.dates = NULL,
  change.factor.precip.mean = NULL,
  change.factor.precip.variance = NULL,
  change.factor.temp.mean = NULL,
  transient.temp.change = TRUE,
  transient.precip.change = TRUE,
  calculate.pet = TRUE,
  compute.parallel = FALSE,
  num.cores = NULL,
  fit.method = "mme"
)
}
\arguments{
\item{climate.data}{placeholder}

\item{climate.grid}{placeholder}

\item{sim.dates}{placeholder}

\item{change.factor.precip.mean}{placeholder}

\item{change.factor.precip.variance}{placeholder}

\item{change.factor.temp.mean}{placeholder}

\item{transient.temp.change}{logical value indicating whether temperature changes are applied gradually (transient) or stepwise}

\item{transient.precip.change}{logical value indicating whether precipitation changes are applied gradually (transient) or stepwise}

\item{calculate.pet}{shoud pet be calculated? (logical)}

\item{compute.parallel}{should the function run in parallel mode (logical)}

\item{num.cores}{number of cores reserved for parallel computing. Default value is maximum cores minus one (omit if compute.parallel is FALSE)}

\item{fit.method}{placeholder}
}
\description{
Perturb weather realizations to reflect climate change
}
