% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeClimateChanges.R
\name{imposeClimateChanges}
\alias{imposeClimateChanges}
\title{Climate Change perturbations function}
\usage{
imposeClimateChanges(
  input.data = NULL,
  coordGrid = NULL,
  file.suffix = "",
  file.prefix = "clim_change_rlz",
  output.path = NULL,
  sim.date.start = NULL,
  variables = NULL,
  variable.units = NULL,
  nc.dimnames = NULL,
  change.settings = NULL,
  save.scenario.matrix = TRUE,
  step = TRUE
)
}
\arguments{
\item{input.data}{Placeholder.}

\item{file.suffix}{Placeholder.}

\item{output.path}{To be completed...}

\item{sim.date.start}{To be completed...}

\item{variables}{To be completed...}

\item{variable.units}{To be completed...}

\item{nc.dimnames}{To be completed...}

\item{change.settings}{To be completed...}

\item{save.scenario.matrix}{Placeholder.}

\item{input.path}{Placeholder.}
}
\value{

}
\description{
Climate Change perturbations function
}
