% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeClimateChanges.R
\name{imposeClimateChanges}
\alias{imposeClimateChanges}
\title{Climate Change perturbations function}
\usage{
imposeClimateChanges(
  proj.name = NULL,
  in.path = NULL,
  in.file = NULL,
  file.suffix = "",
  out.path = NULL,
  sim.date.begin = NULL,
  wg.vars = NULL,
  wg.var.units = NULL,
  nc.dimnames = NULL,
  change.settings = NULL,
  save.scenario.matrix = TRUE
)
}
\arguments{
\item{proj.name}{To be completed...}

\item{in.path}{Placeholder.}

\item{in.file}{Placeholder.}

\item{file.suffix}{Placeholder.}

\item{out.path}{To be completed...}

\item{sim.date.begin}{To be completed...}

\item{wg.vars}{To be completed...}

\item{wg.var.units}{To be completed...}

\item{nc.dimnames}{To be completed...}

\item{change.settings}{To be completed...}

\item{save.scenario.matrix}{Placeholder.}
}
\value{

}
\description{
Climate Change perturbations function
}
