% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnDaily.R
\name{knnDaily}
\alias{knnDaily}
\title{KNN daily sampling}
\usage{
knnDaily(
  cur_sim_PRCP = NULL,
  cur_sim_TEMP = NULL,
  PRCP_TODAY = NULL,
  TEMP_TODAY = NULL,
  k = NULL,
  sd_monthly_PRCP = NULL,
  sd_monthly_TEMP = NULL,
  mean_monthly_PRCP = NULL,
  mean_monthly_TEMP = NULL,
  k1 = NULL,
  count = NULL,
  seed = NULL
)
}
\arguments{
\item{cur_sim_PRCP}{placeholder}

\item{cur_sim_TEMP}{placeholder}

\item{PRCP_TODAY}{placeholder}

\item{TEMP_TODAY}{placeholder}

\item{k}{placeholder}

\item{sd_monthly_PRCP}{placeholder}

\item{sd_monthly_TEMP}{placeholder}

\item{mean_monthly_PRCP}{placeholder}

\item{mean_monthly_TEMP}{placeholder}

\item{k1}{placeholder}

\item{count}{placeholder}

\item{seed}{random seed value}

\item{DATE_TOMORROW}{placeholder}
}
\description{
KNN daily sampling
}
