% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warm_filtering.R
\name{log_filtering_start}
\alias{log_filtering_start}
\title{Log initial setup information}
\usage{
log_filtering_start(
  n_obs,
  n_sim,
  n_realizations,
  sample_target,
  relax_priority
)
}
\description{
Log initial setup information
}
\keyword{internal}
