% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warm_filtering.R
\name{log_filtering_start}
\alias{log_filtering_start}
\title{Log initial setup information}
\usage{
log_filtering_start(
  n_obs,
  n_sim,
  n_realizations,
  sample_target,
  relax_priority
)
}
\arguments{
\item{n_obs}{Integer scalar. Number of observations in observed series.}

\item{n_sim}{Integer scalar. Number of years in simulated series used after windowing.}

\item{n_realizations}{Integer scalar. Number of candidate realizations.}

\item{sample_target}{Integer scalar. Target number to select.}

\item{relax_priority}{Character vector. Relaxation priority vector.}
}
\value{
Invisibly returns NULL.
}
\description{
Internal helper. Displays general information at the start of filtering.
}
\keyword{internal}
