% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warm_filtering.R
\name{log_final_summary}
\alias{log_final_summary}
\title{Log final filtering summary}
\usage{
log_final_summary(pool_size, n_total, n_sampled, relaxation_level)
}
\description{
Log final filtering summary
}
\keyword{internal}
