% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{make_noleap_dates}
\alias{make_noleap_dates}
\title{Generate a no-leap (365-day) daily date sequence}
\usage{
make_noleap_dates(start_date, n_dates)
}
\arguments{
\item{start_date}{Date or character coercible to Date.
The first date of the sequence. If the date is February 29, it is mapped to
February 28 because the no-leap template does not include Feb 29.}

\item{n_dates}{Integer. Total number of dates to generate.}
}
\value{
A vector of class \code{Date} with length \code{n_dates}, following
  a no-leap (365-day) calendar.
}
\description{
Constructs a continuous sequence of dates starting from a given start date,
using a no-leap calendar (i.e., February 29 removed). The function repeats a
365-day month-day template and increments the year every 365 days.
}
\examples{
make_noleap_dates("1980-01-01", 365)
make_noleap_dates(as.Date("1995-02-28"), 1000)

}
