% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quantile_mapping.R
\name{plot.qmap_diagnostics}
\alias{plot.qmap_diagnostics}
\title{Plot method for qmap_diagnostics}
\usage{
\method{plot}{qmap_diagnostics}(x, which = c("all", "moments", "quantiles", "extremes", "monthly"), ...)
}
\description{
Plot method for qmap_diagnostics
}
