% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare data for plotting}
\usage{
prepare_plot_data(obs.results, sim.results, variables, n_grids)
}
\description{
Prepare data for plotting
}
\keyword{Prepare}
\keyword{data}
\keyword{for}
\keyword{internal}
\keyword{plotting}
