% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_diagnostics.R
\name{print.prcp_qm_diagnostics}
\alias{print.prcp_qm_diagnostics}
\title{Print method for precipitation QM diagnostics}
\usage{
\method{print}{prcp_qm_diagnostics}(x, ...)

\method{print}{prcp_qm_diagnostics}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"prcp_qm_diagnostics"}.}

\item{...}{Not used. Included for S3 method consistency.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print method for precipitation QM diagnostics

Print precipitation QM diagnostics
}
