% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{print.prcp_qm_diagnostics}
\alias{print.prcp_qm_diagnostics}
\title{Print precipitation QM diagnostics}
\usage{
\method{print}{prcp_qm_diagnostics}(x, ...)
}
\description{
Print precipitation QM diagnostics
}
