% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{process_observed_data}
\alias{process_observed_data}
\title{Process observed weather data}
\usage{
process_observed_data(obs_data, vars, n_grid, wet_q, extreme_q)
}
\description{
Process observed weather data
}
\keyword{internal}
