% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_weather_generator.R
\name{process_simulated_data}
\alias{process_simulated_data}
\title{Process simulated weather data}
\usage{
process_simulated_data(daily.sim, realization.num, variables, mc.thresholds)
}
\description{
Process simulated weather data
}
\keyword{internal}
