% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileMapping.R
\name{quantileMapping}
\alias{quantileMapping}
\title{Function for perturbing climate statistics using Quantile-Mapping}
\usage{
quantileMapping(
  value = NULL,
  date = NULL,
  par = NULL,
  mon.ts = NULL,
  year.ts = NULL
)
}
\arguments{
\item{value}{To be completed...}

\item{date}{To be completed...}

\item{par}{List of parameters to be passed.}

\item{operator}{A character string to specify the type of perturb change operator.}
}
\value{

}
\description{
Function for perturbing climate statistics using Quantile-Mapping
}
