% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNetcdf.R
\name{readNetcdf}
\alias{readNetcdf}
\title{Read netcdf files}
\usage{
readNetcdf(
  nc.file = NULL,
  nc.dimnames = NULL,
  nc.variables = NULL,
  origin.date = NULL,
  has.leap.days = TRUE,
  omit.empty.cells = TRUE
)
}
\arguments{
\item{nc.file}{The name of the netcdf file to be read}

\item{nc.dimnames}{Desired dimension names in the output data}

\item{nc.variables}{Desired variable names to be read from the netcdf file}

\item{origin.date}{A character string to specify the date of origin}

\item{has.leap.days}{A logical value indicating whether the gridded data includes leap days.}

\item{omit.empty.cells}{A lotical value indicating whether empty cells with missing values to be removed}
}
\value{

}
\description{
Wrapper function to read files and create list object that returns a list of weather data for each grid,
a table of grid coordinates and a date series as well as raw data, dimensions, variables, and attributes
from the netcdf file.
}
\examples{
\dontrun{
output <- readNetcdf(nc.path = system.file('extdata', package = 'gridwegen'),
    nc.file = "ntoum.nc", nc.dimnames = list(x = "lon", y = "lat", time = "time"),
    nc.variables = c("precip", "temp", "temp_min", "temp_max"),
    origin.date = as.Date("1981-01-01"),
    has.leap.days = TRUE)
}
}
