% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNetcdf.R
\name{readNetcdf}
\alias{readNetcdf}
\title{Read netcdf files}
\usage{
readNetcdf(
  nc.file = NULL,
  leap.days = TRUE,
  omit.empty = TRUE,
  spatial.ref = "spatial_ref"
)
}
\arguments{
\item{nc.file}{The name of the netcdf file to be read}

\item{leap.days}{A logical value indicating whether the gridded data includes leap days.}

\item{omit.empty}{A lotical value indicating whether empty cells with missing values to be removed}

\item{spatial.ref}{PLACEHOLDER}
}
\description{
Wrapper function to read files and create list object that returns a list of weather data for each grid,
a table of grid coordinates and a date series as well as raw data, dimensions, variables, and attributes
from the netcdf file.
}
\examples{
\dontrun{
output <- readNetcdf(nc.path = system.file('extdata', package = 'gridwegen'),
    nc.file = "ntoum.nc", nc_dimnames = list(x = "lon", y = "lat", time = "time"),
    nc.variables = c("precip", "temp", "temp_min", "temp_max"),
    origin_date = as.Date("1981-01-01"),
    leap.days = TRUE)
}
}
