% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNetcdf.R
\name{readNetcdf}
\alias{readNetcdf}
\title{Read a gridded NetCDF file and return weather data in tidy format}
\usage{
readNetcdf(
  nc.file,
  leap.days = TRUE,
  omit.empty = TRUE,
  spatial.ref = "spatial_ref",
  signif.digits = NULL
)
}
\arguments{
\item{nc.file}{Path to the NetCDF file.}

\item{leap.days}{Logical: does data include leap days? (default: TRUE)}

\item{omit.empty}{Logical: remove empty/missing-value cells? (default: TRUE)}

\item{spatial.ref}{String: name of spatial reference variable in NetCDF (default: "spatial_ref")}

\item{signif.digits}{(optional) Number of significant digits to round data.}
}
\value{
List with elements:
  \item{data}{List of tibbles of weather variables for each grid cell (length = number of cells)}
  \item{grid}{Tibble of grid coordinates and indices}
  \item{date}{Vector of Date objects}
  \item{dimensions}{Named list of dimension variables (time, x, y, etc.)}
  \item{attributes}{List of attributes for variables and global attributes}
}
\description{
Read a gridded NetCDF file and return weather data in tidy format
}
