% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleDates.R
\name{resampleDates}
\alias{resampleDates}
\title{Dissaggregation function}
\usage{
resampleDates(
  PRCP_FINAL_ANNUAL_SIM = NULL,
  ANNUAL_PRCP = NULL,
  PRCP = NULL,
  TEMP = NULL,
  water_year_start = NULL,
  water_year_end = NULL,
  START_YEAR_SIM = NULL,
  WATER_YEAR_A = NULL,
  k1 = NULL,
  ymax = NULL,
  dates.d = NULL,
  sim.dates.d = NULL,
  YEAR_D = NULL,
  month_list = NULL,
  kk = NULL,
  knn.annual.sample.num = 20,
  thresh1 = 0.3,
  extreme_quantile = 0.8
)
}
\arguments{
\item{PRCP_FINAL_ANNUAL_SIM}{placeholder}

\item{ANNUAL_PRCP}{placeholder}

\item{PRCP}{placeholder}

\item{TEMP}{placeholder}

\item{water_year_start}{placeholder}

\item{water_year_end}{placeholder}

\item{WATER_YEAR_A}{placeholder}

\item{k1}{placeholder}

\item{ymax}{placeholder}

\item{YEAR_D}{placeholder}

\item{month_list}{placeholder}

\item{knn.annual.sample.num}{placeholder}

\item{WATER_YEAR_D}{placeholder}

\item{DATE_D}{placeholder}

\item{MONTH_D}{placeholder}

\item{MONTH_DAY_D}{placeholder}

\item{SIM_LENGTH}{placeholder}
}
\value{

}
\description{
Dissaggregation function
}
