% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleDates.R
\name{resampleDates}
\alias{resampleDates}
\title{Dissaggregation function}
\usage{
resampleDates(
  PRCP_FINAL_ANNUAL_SIM = NULL,
  ANNUAL_PRCP = NULL,
  PRCP = NULL,
  TEMP = NULL,
  START_YEAR_SIM = NULL,
  k1 = NULL,
  ymax = NULL,
  dates.d = NULL,
  sim.dates.d = NULL,
  YEAR_D = NULL,
  month.start = NULL,
  kk = NULL,
  knn.annual.sample.num = 50,
  thresh1 = 0.3,
  extreme_quantile = 0.8
)
}
\arguments{
\item{PRCP_FINAL_ANNUAL_SIM}{placeholder}

\item{ANNUAL_PRCP}{placeholder}

\item{PRCP}{placeholder}

\item{TEMP}{placeholder}

\item{k1}{placeholder}

\item{ymax}{placeholder}

\item{YEAR_D}{placeholder}

\item{knn.annual.sample.num}{placeholder}
}
\value{

}
\description{
Dissaggregation function
}
