% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateWeather.R
\name{simulateWeather}
\alias{simulateWeather}
\title{Simulate gridded weather function}
\usage{
simulateWeather(
  climate.data = NULL,
  climate.grid = NULL,
  year.start = NULL,
  year.num = NULL,
  month.start = 1,
  variable.names = NULL,
  variable.labels = NULL,
  variable.units = NULL,
  sim.year.start = 2020,
  sim.year.num = 40,
  realization.num = 5,
  warm.variable = "precip",
  warm.signif.level = 0.9,
  warm.sample.size = 10000,
  save.warm.results = TRUE,
  knn.annual.sample.size = 50,
  evaluate.model = FALSE,
  evaluate.grid.num = 20,
  apply.delta.changes = TRUE,
  apply.step.changes = TRUE,
  delta.precip = NULL,
  delta.temp = NULL,
  save.scenario.matrix = TRUE,
  output.path = NULL,
  output.ncfile.template = NULL,
  output.ncfile.prefix = "clim_change_rlz",
  ...
)
}
\arguments{
\item{climate.data}{list of data frames of daily weather observations per grid cell. Each data frame, columns are weather variables and rows are daily values.}

\item{climate.grid}{Data frame of grid cells. Each grid cell is assigned an id starting from 1, x and y coordinate index value, and x and y coordinates.}

\item{year.start}{numeric value indicating the desired starting year of the historical data as input}

\item{year.num}{numeric value indicating the total number of years of historical data to be considered as input.}

\item{month.start}{the first month of the year (default value is 1). Use a value other than 1 for water-year based analyses}

\item{variable.names}{vector of names for the variables to be included in the weather generator}

\item{variable.labels}{vector of labels for the weather variables (optional). If no values provided, it is labels will be same as the names}

\item{variable.units}{vector of units for each of the weather variables (optional). If no values provided, a blank vector is used.}

\item{sim.year.start}{numeric value indicating the starting year of the generated time-series}

\item{sim.year.num}{numeric value indicating the desired total number of years of simulated weather realizations}

\item{realization.num}{number of natural variability realizations to be generated.}

\item{warm.variable}{the name of the variable for the wavelet autoregressive mode. Default value is precipitation variable.}

\item{warm.signif.level}{the significance level for the warm model.}

\item{warm.sample.size}{number of annual sequeces to be generated from the the warm model}

\item{save.warm.results}{save warm results to file}

\item{knn.annual.sample.size}{number of knn years to be sampled}

\item{evaluate.model}{logical value indicating wether to save model evaluation plots}

\item{evaluate.grid.num}{Number of grid cells to be sampled in the evaluation plots}

\item{apply.delta.changes}{Logical value indicating weather delta factors to be applied}

\item{apply.step.changes}{logical value to apply transient or step changes}

\item{delta.precip}{list of precipitation changes}

\item{delta.temp}{List of delta changes}

\item{save.scenario.matrix}{save scenario matrix files}

\item{output.path}{output path for the weather generator results (string)}

\item{output.ncfile.template}{template file for the output netcdf}

\item{output.ncfile.prefix}{the prefix string to be added to resulting netcdf files}

\item{...}{placeholder}
}
\value{

}
\description{
Description goes here....
}
