% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateWeather.R
\name{simulateWeather}
\alias{simulateWeather}
\title{Main wrapper function for daily weather generation}
\usage{
simulateWeather(
  proj.name = NULL,
  output.dir = NULL,
  climate_tidy = NULL,
  wg.date.begin = NULL,
  wg.vars = NULL,
  wg.var.labs = NULL,
  wg.var.units = NULL,
  warm.var = NULL,
  ssig = 0.9,
  ymax = 40,
  rmax = 20000,
  nmax = 5,
  nc.dimnames = NULL,
  validate = FALSE,
  ...
)
}
\arguments{
\item{proj.name}{Placeholder.}

\item{output.dir}{}

\item{wg.date.begin}{Placeholder.}

\item{wg.vars}{Placeholder.}

\item{wg.var.labs}{Placeholder.}

\item{wg.var.units}{Placeholder.}

\item{warm.var}{Placeholder.}

\item{ssig}{Placeholder.}

\item{ymax}{Placeholder.}

\item{rmax}{Placeholder.}

\item{nmax}{Placeholder.}

\item{nc.dimnames}{Placeholder.}

\item{validate}{Placeholder.}

\item{nc.path}{Placeholder.}

\item{nc.file}{Placeholder.}
}
\value{

}
\description{
Main wrapper function for daily weather generation
}
