% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateWeather.R
\name{simulateWeather}
\alias{simulateWeather}
\title{Main wrapper function for daily weather generation}
\usage{
simulateWeather(
  output.path = NULL,
  hist.climate = NULL,
  grid.coords = NULL,
  hist.date.start = NULL,
  variables = NULL,
  variable.labels = NULL,
  variable.units = NULL,
  warm.variable = "precip",
  warm.signif.level = 0.9,
  warm.sample.num = 20000,
  knn.annual.sample.num = 20,
  sim.year.num = 40,
  sim.year.start = 2020,
  realization.num = 5,
  check.stats = FALSE,
  month.start = 6,
  check.grid.num = 50,
  return.date.indices = TRUE,
  save.to.netcdf = FALSE,
  ...
)
}
\arguments{
\item{output.path}{Placeholder.}

\item{hist.climate}{Placeholder.}

\item{hist.date.start}{Placeholder.}

\item{variables}{Placeholder.}

\item{variable.labels}{Placeholder.}

\item{variable.units}{Placeholder.}

\item{warm.variable}{Placeholder.}

\item{warm.signif.level}{Placeholder.}

\item{warm.sample.num}{Placeholder.}

\item{sim.year.num}{Placeholder.}

\item{sim.year.start}{Placeholder.}

\item{realization.num}{Placeholder.}

\item{check.stats}{Placeholder.}

\item{month.start}{Placeholder.}

\item{check.grid.num}{Placeholder.}

\item{...}{Placeholder.}
}
\value{

}
\description{
Main wrapper function for daily weather generation
}
