% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateWeather.R
\name{simulateWeather}
\alias{simulateWeather}
\title{Generate daily, multivariate griddded weather series}
\usage{
simulateWeather(
  output.path = NULL,
  hist.climate = NULL,
  grid.coords = NULL,
  hist.year.start = NULL,
  hist.year.num = NULL,
  variables = NULL,
  variable.labels = NULL,
  variable.units = NULL,
  warm.variable = "precip",
  warm.signif.level = 0.9,
  warm.sample.num = 20000,
  knn.annual.sample.num = 20,
  sim.year.num = 40,
  sim.year.start = 2020,
  realization.num = 5,
  check.stats = FALSE,
  month.start = 5,
  check.grid.num = 50,
  return.date.indices = TRUE,
  save.to.netcdf = FALSE,
  ...
)
}
\arguments{
\item{output.path}{path to save the results}

\item{hist.climate}{list of data frames for historical daily weather data with variable}

\item{grid.coords}{Placeholder}

\item{variables}{Names of weather variables to be included, e.g., precip, temp, temp_min, temp_max.}

\item{variable.labels}{Long names of weather variables. If not provided, same as variable names.}

\item{variable.units}{Measurment units of the weather variables}

\item{warm.variable}{Variable to be used for the wavelet analysis}

\item{warm.signif.level}{Statistical significance level used in the wavelet analysis.}

\item{warm.sample.num}{Placeholder.}

\item{knn.annual.sample.num}{Placeholder}

\item{sim.year.num}{Placeholder.}

\item{sim.year.start}{Placeholder.}

\item{realization.num}{Placeholder.}

\item{check.stats}{Placeholder.}

\item{month.start}{Placeholder.}

\item{check.grid.num}{Placeholder.}

\item{return.date.indices}{Placeholder}

\item{save.to.netcdf}{Placeholder}

\item{...}{Placeholder}

\item{hist.date.start}{Starting year for the historical data to be used in the analysis, e.g., 1980}

\item{hist.date.end}{Ending year for the historical data to be used in the analysis, e.g., 2000}
}
\value{

}
\description{
Generate daily, multivariate griddded weather series
}
