% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateWeather.R
\name{simulateWeather}
\alias{simulateWeather}
\title{Main wrapper function for daily weather generation}
\usage{
simulateWeather(
  output.path = NULL,
  hist.climate = NULL,
  hist.date.start = NULL,
  variables = NULL,
  variable.labels = NULL,
  variable.units = NULL,
  warm.variable = NULL,
  warm.signif.level = 0.9,
  knn.annual.sample.size = 20,
  ymax = 40,
  rmax = 5000,
  nmax = 5,
  validate = TRUE,
  sim.year.start = 2020,
  month.list = 1:12,
  validation.grid.num = 50,
  return.sampled.date.indices = TRUE,
  save.to.netcdf = FALSE,
  ...
)
}
\arguments{
\item{output.path}{Placeholder.}

\item{hist.climate}{Placeholder.}

\item{hist.date.start}{Placeholder.}

\item{variables}{Placeholder.}

\item{variable.labels}{Placeholder.}

\item{variable.units}{Placeholder.}

\item{warm.variable}{Placeholder.}

\item{warm.signif.level}{Placeholder.}

\item{ymax}{Placeholder.}

\item{rmax}{Placeholder.}

\item{nmax}{Placeholder.}

\item{validate}{Placeholder.}

\item{sim.year.start}{Placeholder.}

\item{month.list}{Placeholder.}

\item{validation.grid.num}{Placeholder.}

\item{...}{Placeholder.}
}
\value{

}
\description{
Main wrapper function for daily weather generation
}
