% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_generator.R
\name{standardize_obs_sim_periods}
\alias{standardize_obs_sim_periods}
\title{Standardize obs and sim to same full-year length via random windowing
Applies the chosen window consistently across all grids and all realizations.}
\usage{
standardize_obs_sim_periods(daily.obs, daily.sim, realization.num, variables)
}
\description{
Standardize obs and sim to same full-year length via random windowing
Applies the chosen window consistently across all grids and all realizations.
}
\keyword{internal}
