% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap_diagnostics.R
\name{summary.prcp_qm_diagnostics}
\alias{summary.prcp_qm_diagnostics}
\title{Summarize precipitation QM diagnostics}
\usage{
\method{summary}{prcp_qm_diagnostics}(object, ...)
}
\description{
Summarize precipitation QM diagnostics
}
