% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quantile_mapping.R
\name{summary.qmap_diagnostics}
\alias{summary.qmap_diagnostics}
\title{Summary method for qmap_diagnostics}
\usage{
\method{summary}{qmap_diagnostics}(object, ...)
}
\description{
Summary method for qmap_diagnostics
}
