% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletARSubset.R
\name{waveletARSubset}
\alias{waveletARSubset}
\title{Resample from WARM outputs}
\usage{
waveletARSubset(
  series.obs = NULL,
  series.sim = NULL,
  power.obs = NULL,
  power.sim = NULL,
  power.period = NULL,
  power.signif = NULL,
  sample.num = 5,
  seed = NULL,
  save.plots = TRUE,
  save.series = TRUE,
  output.path = NULL,
  padding = TRUE,
  bounds = list(mean = 0.1, sd = 0.2, min = 0.2, max = 0.2, signif.threshold = 0.5,
    nonsignif.threshold = 1.5)
)
}
\arguments{
\item{series.obs}{A numeric vector of observd time-series values.}

\item{series.sim}{A numeric matrix, with simulated time-series.}

\item{power.obs}{A numeric vector of power spectra of observed time-series.}

\item{power.sim}{A numeric matrix of power spectrum of simulated time-series.}

\item{power.period}{A time-series of power periods calculated.}

\item{power.signif}{A time-series of power significance.}

\item{sample.num}{A numeric value to define the final sample size.}

\item{seed}{A numeric value to define a seed for resampling.}

\item{save.plots}{A logical, to save the plots to file.}

\item{save.series}{A logical to write the results to csv files.}

\item{output.path}{Output folder path}

\item{padding}{placeholder}

\item{bounds}{placeholder}
}
\description{
Resample from WARM outputs
}
