% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletARSubset.R
\name{waveletARSubset}
\alias{waveletARSubset}
\title{Resample from WARM outputs}
\usage{
waveletARSubset(
  series.obs = NULL,
  series.sim = NULL,
  power.obs = NULL,
  power.sim = NULL,
  power.period = NULL,
  power.signif = NULL,
  sample.num = 5,
  seed = NULL,
  save.plots = TRUE,
  save.series = TRUE,
  verbose = FALSE,
  output.path = NULL,
  padding = TRUE,
  bounds = list(mean = c(0.95, 1.05), sd = c(0.9, 1.1), min = c(0.9, 1.1), max = c(0.9,
    1.1), power = c(0.6, 2.25), nonsignif.threshold = 0.9)
)
}
\arguments{
\item{series.obs}{A numeric vector of observd time-series values.}

\item{series.sim}{A numeric matrix, with simulated time-series.}

\item{power.obs}{A numeric vector of power spectra of observed time-series.}

\item{power.sim}{A numeric matrix of power spectrum of simulated time-series.}

\item{power.period}{A time-series of power periods calculated.}

\item{power.signif}{A time-series of power significance.}

\item{sample.num}{A numeric value to define the final sample size.}

\item{seed}{A numeric value to define a seed for resampling.}

\item{save.plots}{A logical, to save the plots to file.}

\item{save.series}{A logical to write the results to csv files.}

\item{verbose}{A logical to decide if further information to be displayed on the screen.}

\item{output.path}{Output folder path}
}
\value{

}
\description{
Resample from WARM outputs
}
