% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletARSubset.R
\name{waveletARSubset}
\alias{waveletARSubset}
\title{A function to resample from WARM model results}
\usage{
waveletARSubset(
  series.obs = NULL,
  series.sim = NULL,
  power.obs = NULL,
  power.sim = NULL,
  power.period = NULL,
  power.signif = NULL,
  nmax = NULL,
  seed = NULL,
  save.plots = TRUE,
  save.series = TRUE,
  verbose = FALSE,
  out.path = out_path,
  mean.bounds = c(0.95, 1.05),
  sdev.bounds = c(0.9, 1.1),
  max.bounds = c(0.9, 1.1),
  min.bounds = c(0.9, 1.1),
  power.bounds = c(0.75, 2.5),
  nonsig.threshold = 0.8
)
}
\arguments{
\item{series.obs}{A numeric vector of observd time-series values.}

\item{series.sim}{A numeric matrix, with simulated time-series.}

\item{power.obs}{A numeric vector of power spectra of observed time-series.}

\item{power.sim}{A numeric matrix of power spectrum of simulated time-series.}

\item{power.period}{A time-series of power periods calculated.}

\item{power.signif}{A time-series of power significance.}

\item{nmax}{A numeric value to define the final sample size.}

\item{seed}{A numeric value to define a seed for resampling.}

\item{save.plots}{A logical, to save the plots to file.}

\item{save.series}{A logical to write the results to csv files.}

\item{verbose}{A logical to decide if further information to be displayed on the screen.}

\item{out.path}{Output folder path}

\item{mean.bounds}{A numeric vector, defining the minimum and maximum limits for the mean value of the time-series when resampling.}

\item{sdev.bounds}{A numeric vector, defining the minimum and maximum limits for the standard deviation of the time-series when resampling.}

\item{max.bounds}{A numeric vector, defining the minimum and maximum limits for the maximum value of the time-series when resampling.}

\item{min.bounds}{A numeric vector, defining the minimum and maximum limits for the minimum value of the time-series when resampling.}

\item{power.bounds}{A numeric vector, defining the minimum and maximum limits for power spectra.}

\item{nonsig.threshold}{A numeric vector to define a resampling threshold for sampling.}
}
\value{

}
\description{
A function to resample from WARM model results
}
