% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletDecompose.R
\name{waveletDecompose}
\alias{waveletDecompose}
\title{Function for weather generator decomposition}
\usage{
waveletDecompose(
  variable = NULL,
  signif.periods = NULL,
  noise.type = "white",
  signif.level = 0.8,
  plot = TRUE,
  output.path = NULL
)
}
\arguments{
\item{variable}{A vector of time-series of weather variables.}

\item{signif.periods}{Significant low-frequency periods in the original time-series.}

\item{noise.type}{A logical specifying the type of background noise.}

\item{signif.level}{Significance level for the wavelet analysis.}

\item{plot}{Draw plot}

\item{output.path}{Output path}
}
\value{

}
\description{
Function for weather generator decomposition
}
