% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNetcdf.R
\name{writeNetcdf}
\alias{writeNetcdf}
\title{Title}
\usage{
writeNetcdf(
  data = NULL,
  coord.grid = NULL,
  output.path = NULL,
  nc.dimensions = NULL,
  nc.dimnames = NULL,
  origin.date = NULL,
  calendar.type = "no leap",
  variables = NULL,
  variable.units = NULL,
  nc.compression = 4,
  file.prefix = "clim_change_rlz",
  file.suffix = ""
)
}
\arguments{
\item{output.path}{Placeholder}

\item{nc.dimensions}{Placeholder}

\item{nc.dimnames}{Placeholder}

\item{origin.date}{Placeholder}

\item{calendar.type}{Placeholder}

\item{variables}{Placeholder}

\item{variable.units}{Placeholder}

\item{nc.compression}{Placeholder}

\item{sim.length}{Placeholder}
}
\value{

}
\description{
Title
}
