% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNetcdf.R
\name{writeNetcdf}
\alias{writeNetcdf}
\title{Write to netcdf file}
\usage{
writeNetcdf(
  data = NULL,
  coord.grid = NULL,
  output.path = NULL,
  origin.date = NULL,
  calendar.type = "noleap",
  nc.template.file = NULL,
  nc.compression = 4,
  nc.spatial.ref = "spatial_ref",
  nc.file.prefix = "clim_change_rlz",
  nc.file.suffix = ""
)
}
\arguments{
\item{data}{placeholder}

\item{coord.grid}{placeholder}

\item{output.path}{Placeholder}

\item{origin.date}{Placeholder}

\item{calendar.type}{Placeholder}

\item{nc.template.file}{placeholder}

\item{nc.compression}{Placeholder}

\item{nc.file.prefix}{placeholder}

\item{nc.file.suffix}{placeholder}

\item{variables}{Placeholder}

\item{variable.units}{Placeholder}
}
\value{

}
\description{
Write to netcdf file
}
