% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNetcdf.R
\name{writeNetcdf}
\alias{writeNetcdf}
\title{Write to netcdf file}
\usage{
writeNetcdf(
  nc.temp = NULL,
  data = NULL,
  coord.grid = NULL,
  output.path = NULL,
  nc.dimnames = NULL,
  origin.date = NULL,
  calendar.type = "no leap",
  variables = NULL,
  variable.units = NULL,
  nc.compression = 4,
  file.prefix = "clim_change_rlz",
  file.suffix = ""
)
}
\arguments{
\item{nc.temp}{placeholder}

\item{data}{placeholder}

\item{coord.grid}{placeholder}

\item{output.path}{Placeholder}

\item{nc.dimnames}{Placeholder}

\item{origin.date}{Placeholder}

\item{calendar.type}{Placeholder}

\item{variables}{Placeholder}

\item{variable.units}{Placeholder}

\item{nc.compression}{Placeholder}

\item{file.prefix}{placeholder}

\item{file.suffix}{placeholder}
}
\value{

}
\description{
Write to netcdf file
}
